### 2019-09-02 #50 @whedon
[PRE REVIEW]: ModelingToolkit.jl: An Intermediate Representation for Scientific Domain-Specific Languages
**Submitting author:** @HarrisonGrodin (<a href="http://orcid.org/0000-0002-0947-3520">Harrison Grodin</a>)
**Repository:** <a href="https://github.com/JuliaDiffEq/ModelingToolkit.jl" target ="_blank">https://github.com/JuliaDiffEq/ModelingToolkit.jl</a>

Thanks for submitting your paper to JuliaCon @HarrisonGrodin. The JuliaCon editor @matbesancon, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@HarrisonGrodin if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
