### 2019-09-04 #50 @matbesancon
Hi @HarrisonGrodin, for the juliacon proceedings we prefer having the paper in the same repo as the code. If you don't want to pollute the repo you can put it in another branch
