+++ test/runtests.jl
@@ -5,3 +5,4 @@ using ModelingToolkit, Test
 @testset "Simplify Test" begin include("simplify.jl") end
 @testset "Direct Usage Test" begin include("direct.jl") end
 @testset "System Construction Test" begin include("system_construction.jl") end
+@testset "Distributed Test" begin include("distributed.jl") end
