+++ paper/paper.tex
@@ -1,0 +1,309 @@
+% JuliaCon proceedings template
+\documentclass{juliacon}
+\setcounter{page}{1}
+
+\usepackage{amsmath}
+\usepackage{tikz}
+
+\lstdefinelanguage{Julia}%
+  {morekeywords={abstract,break,case,catch,const,continue,do,else,elseif,%
+      end,export,false,for,function,immutable,import,importall,if,in,%
+      macro,module,otherwise,quote,return,struct,switch,true,try,typealias,%
+      using,while},%
+   sensitive=true,%
+   alsoother={\$},%
+   morecomment=[l]\#,%
+   morecomment=[n]{\#=}{=\#},%
+   morestring=[s]{"}{"},%
+   morestring=[m]{'}{'},%
+}[keywords,comments,strings]%
+
+
+\lstset{
+    language = Julia,
+    frame = single
+}
+
+\begin{document}
+
+\input{header}
+
+\maketitle
+
+\begin{abstract}
+
+Domain-specific languages (DSLs) enable scientific programmers to describe their models in a high-level format while still allowing for internal optimizations. However, while the homoiconicity of Julia has encouraged the proliferation of DSLs, code reuse between DSLs has been stymied due to the lack of a common intermediate representation for performing inspection, symbolic transformation, and compilation of scientific models. In this paper, we present ModelingToolkit.jl, a Julia library and intermediate representation for scientific DSLs. We discuss the internal representations of symbolic expressions and systems, chosen to implicitly enforce logical invariants. We showcase how ModelingToolkit is being used in the pharmacometric modeling DSL Pumas.jl, and describe future improvements to incorporate this system into other Julia packages.
+
+\headingtable
+
+\end{abstract}
+
+\section{Introduction}
+
+In many applications, scientists and engineers need to perform symbolic transformations on their models in order to have a formulation that is efficiently computable by numerical methods. For example, Modelica \cite{mattsson_modelica_1997} is a language for describing physical differential-algebraic equations (DAEs) and utilizes the Pantelides algorithm \cite{doi:10.1137/0909014} to transform higher index systems to index 1 DAEs, which can be handled by numerical solvers like DASSL \cite{petzold_description_nodate} or the IDA method of SUNDIALS \cite{alan_c._hindmarsh_sundials:_2005}. NONMEM \cite{sheiner1977estimation} is a domain-specific language for pharmacometrics in Fortran that reads a text file describing differential equation models and appends the sensitivity equations utilized by parameter estimation routines. These packages are just two among many which have developed their own internal systems for performing the steps necessary to handle this process, which includes:
+
+\begin{itemize}
+    \item Parsing a text file into an internal representation (IR).
+    \item Performing transformations on the IR to build new model expressions.
+    \item Compiling the expressions to a common format to be utilized by numerical methods.
+\end{itemize}
+
+While the distinct tools achieve the goals of the end user, the internal mathematical systems for computing the model transformations cannot be utilized outside of their respective frameworks.
+
[…]
