+++ paper/journal_dat.tex
@@ -1,0 +1,6 @@
+% **************GENERATED FILE, DO NOT EDIT**************
+
+\def\@journalName{Proceedings of JuliaCon}
+\def\@volume{1}
+\def\@issue{1}
+\def\@year{2019}
