+++ paper/header.tex
@@ -1,0 +1,12 @@
+% **************GENERATED FILE, DO NOT EDIT**************
+
+\title{ModelingToolkit.jl: An Intermediate Representation for Scientific Domain-Specific Languages}
+
+\author[1]{Harrison Grodin}
+\author[2]{Yingbo Ma}
+\author[3]{Christopher Rackauckas}
+\affil[1]{School of Computer Science, Carnegie Mellon University}
+\affil[2]{Department of Mathematics, University of California Irvine}
+\affil[3]{Department of Mathematics, Massachusetts Institute of Technology}
+
+\keywords{Julia, symbolic computation, differential equations, pharmacometrics}
