+++ paper/ref.bib
@@ -1,0 +1,467 @@
+@article{bezanson_julia:_2017,
+	title = {Julia: {A} {Fresh} {Approach} to {Numerical} {Computing}},
+	volume = {59},
+	issn = {0036-1445},
+	doi = {10.1137/141000671},
+	journal = {SIAM Review},
+	author = {Bezanson, J. and Edelman, A. and Karpinski, S. and Shah, V.},
+	month = jan,
+	year = {2017},
+	pages = {65--98},
+	annote = {The following values have no corresponding Zotero field:number: 1publisher: Society for Industrial and Applied Mathematics}
+}
+
+@inproceedings{Otter2017TransformationOD,
+  title={Transformation of Differential Algebraic Array Equations to Index One Form},
+  author={Martin Otter and Hilding Elmqvist},
+  year={2017}
+}
+
+@article{mogensen2018optim,
+  author  = {Mogensen, Patrick Kofod and Riseth, Asbj{\o}rn Nilsen},
+  title   = {Optim: A mathematical optimization package for {Julia}},
+  journal = {Journal of Open Source Software},
+  year    = {2018},
+  volume  = {3},
+  number  = {24},
+  pages   = {615},
+  doi     = {10.21105/joss.00615}
+}
+
+@misc{diffeqbiological,
+	title = {DiffEqBiological.jl},
+	url = {https://github.com/JuliaDiffEq/DiffEqBiological.jl},
+	urldate = {2018-4-11},
+	author = {Loman, Torkel and Isaacson, Samuel and Rackauckas, Christopher}
+}
+
+@misc{symengine,
+	title = {SymEngine.jl},
+	url = {https://github.com/symengine/SymEngine.jl},
+	urldate = {2018-4-11},
+	author = {Fernando, Isuru}
+}
+
+
+@article{mould_basic_2013,
+	title = {Basic {Concepts} in {Population} {Modeling}, {Simulation}, and {Model}-{Based} {Drug} {Development}—{Part} 2: {Introduction} to {Pharmacokinetic} {Modeling} {Methods}},
+	volume = {2},
+	issn = {2163-8306},
+	url = {https://doi.org/10.1038/psp.2013.14},
+	doi = {10.1038/psp.2013.14},
+	abstract = {Population pharmacokinetic models are used to describe the time course of drug exposure in patients and to investigate sources of variability in patient exposure. They can be used to simulate alternative dose regimens, allowing for informed assessment of dose regimens before study conduct. This paper is the second in a three-part series, providing an introduction into methods for developing and evaluating population pharmacokinetic models. Example model files are available in the Supplementary Data online. CPT: Pharmacometrics  doi:10.1038/psp.2013.14; advance online publication 17 April 2013},
+	number = {4},
+	urldate = {2018-12-01},
+	journal = {CPT: Pharmacometrics \& Systems Pharmacology},
+	author = {Mould, DR and Upton, RN},
+	month = apr,
+	year = {2013},
+	pages = {38},
+	annote = {doi: 10.1038/psp.2013.14}
+}
+
+
+@article{sommer_numerical_nodate,
+	title = {Numerical {Methods} for {Parameter} {Estimation} in {Dynamical} {Systems} with {Noise}},
+	language = {de},
+	author = {Sommer, Andreas},
+	pages = {282},
[…]

[…]
