### 2019-09-05 #50 @HarrisonGrodin
I've added it to the `paper` branch, in the `paper` directory:
https://github.com/JuliaDiffEq/ModelingToolkit.jl/tree/paper
