### 2019-09-05 #50 @matbesancon
@HarrisonGrodin there may be some illegal characters in the bibtex keys?
