### 2019-09-12 209583a
GG -> GeneralizedGenerated
