### 2019-09-13 4735993
fix direct tests
