### 2019-09-13 34e01d0
almost to new GeneratedGeneralized
