+++ .travis.yml
@@ -4,6 +4,7 @@ os:
   - linux
 #  - osx
 julia:
+  - 1.0
   - 1.1
   - nightly
 
