+++ test/distributed.jl
@@ -3,24 +3,23 @@ using Distributed
 addprocs(2)
 
 @everywhere using ModelingToolkit, OrdinaryDiffEq
-using
 
 # create the Lorenz system
-@parameters t σ ρ β
-@variables x(t) y(t) z(t)
-@derivatives D'~t
+@everywhere @parameters t σ ρ β
+@everywhere @variables x(t) y(t) z(t)
+@everywhere @derivatives D'~t
 
-eqs = [D(x) ~ σ*(y-x),
+@everywhere eqs = [D(x) ~ σ*(y-x),
        D(y) ~ x*(ρ-z)-y,
        D(z) ~ x*y - β*z]
 
-de = ODESystem(eqs)
-ode_func = ODEFunction(de, [x,y,z], [σ, ρ, β])
+@everywhere de = ODESystem(eqs)
+@everywhere ode_func = ODEFunction(de, [x,y,z], [σ, ρ, β])
 
-u0 = [19.,20.,50.]
-params = [16.,45.92,4]
+@everywhere u0 = [19.,20.,50.]
+@everywhere params = [16.,45.92,4]
 
-ode_prob = ODEProblem(ode_func, u0, (0., 10.),params)
+@everywhere ode_prob = ODEProblem(ode_func, u0, (0., 10.),params)
 
 @everywhere begin
 
