### 2019-09-19 #52 @whedon
[REVIEW]: ModelingToolkit.jl: An Intermediate Representation for Scientific Domain-Specific Languages
**Submitting author:** @HarrisonGrodin (<a href="http://orcid.org/0000-0002-0947-3520">Harrison Grodin</a>)
**Repository:** <a href="https://github.com/JuliaDiffEq/ModelingToolkit.jl" target ="_blank">https://github.com/JuliaDiffEq/ModelingToolkit.jl</a>

Claudio created [Issue #182](https://github.com/JuliaDiffEq/ModelingToolkit.jl/issues/182)
Claudio created [Issue #184](https://github.com/JuliaDiffEq/ModelingToolkit.jl/issues/184) and [Issue #185](https://github.com/JuliaDiffEq/ModelingToolkit.jl/issues/185)

Claudio created [Issue #186](https://github.com/JuliaDiffEq/ModelingToolkit.jl/issues/186)
Claudio created [Issue #187](https://github.com/JuliaDiffEq/ModelingToolkit.jl/issues/187)
Claudio created [Issue #187](https://github.com/JuliaDiffEq/ModelingToolkit.jl/issues/187)

Claudio: No.
