+++ src/ModelingToolkit.jl
@@ -89,6 +89,7 @@ include("function_registration.jl")
 include("simplify.jl")
 include("utils.jl")
 include("direct.jl")
+include("domains.jl")
 include("systems/diffeqs/diffeqsystem.jl")
 include("systems/diffeqs/first_order_transform.jl")
 include("systems/nonlinear/nonlinear_system.jl")
