+++ src/systems/pde/pdesystem.jl
@@ -1,0 +1,7 @@
+struct PDESystem <: ModelingToolkit.AbstractSystem
+  eq
+  bcs
+  domain
+  indvars
+  depvars
+end
