+++ src/ModelingToolkit.jl
@@ -93,5 +93,6 @@ include("domains.jl")
 include("systems/diffeqs/diffeqsystem.jl")
 include("systems/diffeqs/first_order_transform.jl")
 include("systems/nonlinear/nonlinear_system.jl")
+include("systems/pde/pdesystem.jl")
 
 end # module
