+++ src/domains.jl
@@ -1,3 +1,5 @@
+export IntervalDomain, ProductDomain, ⊗, CircleDomain
+
 abstract type AbstractDomain{T,N} end
 
 struct VarDomainPairing
