+++ Project.toml
@@ -20,9 +20,8 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 julia = "1"
 
 [extras]
-DiffEqOperators = "9fdde737-9c7f-55bf-ade8-46b3f136cc48"
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["DiffEqOperators", "OrdinaryDiffEq", "Test"]
+test = ["OrdinaryDiffEq", "Test"]
