### 2019-09-29 SciML/ModelingToolkit.jl#182 @clagms
Please add installation instructions
Thank you for this work! :)

Could you please add some installation instructions to the README.md?

For someone who might not be experienced in Julia, something like the following would already help.
# Installation
In Julia terminal, type:
```
using Pkg
Pkg.add("ModelingToolkit")
```
