### 2019-09-30 JuliaDiffEq/ModelingToolkit.jl#184 @clagms
Please clarify connection with NLsolve.jl:
It would be nice if, where it is written `We can use this to build a nonlinear function for use with NLsolve.jl:`, in README.md, you could clarify how the resulting function can be used in NLsolve.jl.
The current example code is:
```
f = @eval eval(nlsys_func)
# Make a closure over the parameters for for NLsolve.jl
f2 = (du,u) -> f(du,u,(10.0,26.0,2.33))
```
but it is not clear to me:
- what `u` and `du` are;
- how the resulting function is used in NLsolve; and
- Where is the computation of the Jacobian? I understood that ModelingToolkit.jl uses symbolic techniques to generate the equations for the Jacobian.

The most obvious example usage of NLsolve.jl is:
```
using NLsolve

function f!(F, x)
    F[1] = (x[1]+3)*(x[2]^3-7)+18
    F[2] = sin(x[2]*exp(x[1])-1)
end

function j!(J, x)
    J[1, 1] = x[2]^3-7
    J[1, 2] = 3*x[2]^2*(x[1]+3)
    u = exp(x[1])*cos(x[2]*exp(x[1])-1)
    J[2, 1] = x[2]*u
    J[2, 2] = u
end

nlsolve(f!, j!, [ 0.1; 1.2])
```
