### 2019-09-30 SciML/ModelingToolkit.jl#185 @clagms
Please add example ODE to README.md
It would be nice if an example with ODEs where added to the documentation.

Something like:
```
eqs = [D3(u) ~ 2(D2(u)) +D(u) + D(x) + 1
D2(x) ~ D(x) +2]
de = ODESystem(eqs)
de' = ode_order_lowering(de)
```
and then the example code that:
- computes sensitivity equations
- plugs `de'` to a numerical solver, and computes numerical solution.
This would make it easier for users to recognize the value of this package as a bridge to solve ODEs with local sensitivity.
