### 2019-09-30 SciML/ModelingToolkit.jl#186 @clagms
Please improve motivation in README.md
It would be nice if the advantages of using ModelingToolkit.jl would be clearer at the beginning of README.md.
For example:
```
ModelingToolkit.jl enables an easier implementation of domain specific modelling languages by providing an intermediate representation, and ready to use algorithms that typically improve the performance of the numerical solution to scientific models. Domain specific modelling languages are important because they enable the representation of a problem in the terms that domain experts can understand.
```
