### 2019-09-30 SciML/ModelingToolkit.jl#187 @clagms
Please add instructions on how to contribute
It would be helpful to have some text on how contributions should be made.
This includes how to setup the development environment and how to run the tests.
