### 2019-09-30 #52 @clagms
>Authors: Does the paper.tex file include a list of authors with their affiliations?

I cannot find the repository where the *.tex files are. The pdf satisfies this criteria though.
