### 2019-09-30 #52 @clagms
>- [ ] **Page limit:** Is the page limit for full papers respected by the submitted document?

Where is this limit defined? I cannot find the call for papers. Apologies for my ignorance :P
