### 2019-09-30 #52 @matbesancon
It's all on https://proceedings.juliacon.org/, reviewer and author guide. They respect the page limit, which is not too much of a big deal as long as it remains reasonable
