### 2019-09-30 #52 @clagms
The repository address:
[https://github.com/HarrisonGrodin/JuliaCon2019-Paper](https://github.com/HarrisonGrodin/JuliaCon2019-Paper)
leads to a 404. I've updated the links to
[https://github.com/JuliaDiffEq/ModelingToolkit.jl](https://github.com/JuliaDiffEq/ModelingToolkit.jl)
but I think this was generated from the review process (so the problem is not really solved).
