### 2019-09-30 #52 @clagms
I attach the pdf with the comments on the text:
[10.21105.jcon.00052_review.pdf](https://github.com/JuliaCon/proceedings-review/files/3673234/10.21105.jcon.00052_review.pdf)
