### 2019-10-02 b166fc0
Add basic latexify support for Equation arrays.
