### 2019-10-02 JuliaDiffEq/ModelingToolkit.jl#186 @jamesjscully
I don't understand what the purpose of the Operation type is. Why not just manipulate Expr?

It seems like there are a few disadvantages.
1) littering the global scope
2) You can do AST manipulations with Expr, but you can't really do anything with Operation as far as I can tell.
3) explicitly declaring variables and parameters is not strictly necessary in a lot of cases, and is sort of boiler plate, albeit short. For example you can infer from the equations. ( if D(x) = Expr(...) then x must be a state variable.)

What are the advantages of Operation and how should one interact with it for maximum benefit? It would be nice for this to be explained, because it's not immediately obvious.
