### 2019-10-03 b1cbf01
Add methods to convert Operation to Expr
