### 2019-10-03 3bf6b6c
Update operations.jl
