### 2019-10-03 JuliaDiffEq/ModelingToolkit.jl#186 @ChrisRackauckas
They are completely different things. I would suggest taking a little bit of time to read through the documentation and http://tutorials.juliadiffeq.org/html/ode_extras/01-ModelingToolkit.html and just try to recreate all of those examples with Expr (without type-piracy). When you see where that fails, the idea of Operation is pretty clear.

>littering the global scope

Use `import`?

>You can do AST manipulations with Expr, but you can't really do anything with Operation as far as I can tell.

You can't as easily do AST manipulations with Expr as you can with an Operation, which is exactly the problem...

How do you automatically convert a user's numerical function into a mathematical expression? Compute with operations.

```julia
@parameters t σ ρ β
@variables x(t) y(t) z(t)
function lorenz(du,u,p,t)
 du[1] = p[1]*(u[2]-u[1])
 du[2] = u[1]*(p[2]-u[3]) - u[2]
 du[3] = u[1]*u[2] - p[3]*u[3]
end
u = [x,y,z]
du = similar(u)
p = [σ,ρ,β]
lorenz(du,u,p,t)
du
```

If you wanted to use `Expr`, this would give an `Expr + Expr undefined` type error.

Then, ASTs aren't necessarily a mathematical expression. What is the derivative of a while loop block? The derivative of a begin block? Operation is *desugared* for a reason. That's it's purpose: it's constrained to be a mathematical expression, and it will turn Julia code into a mathematical expression. So doing mathematical manipulations on Operation, like computing the inverse of it's Jacobian, is always something valid to ask. Which is why we can do things like automatically trace a numerical function at runtime to an operation to symbolically differentiate a function: you have to desugar for that.

And because it's a mathematical operation graph, you can actually manipulate it. How do you manipulate it? AbstractTerms.jl/Simplify.jl/Rewrite.jl integration, just changing values in the operation graph, etc. You can't do that with a Julia AST because that's a program AST, so you may not necessarily be computing the same thing (you cannot even assume associativity and such there). Unless you assume in very specific functions that the AST means something very different from Julia IR, that it acts differently, that different constructs are allowed, that there is a different definition of invariance, that functions which may not exist at runtime are allowed to be actors (`D`), etc... but then why are you using the Julia AST if it has a completely different meaning and so if you actually do get an Expr from someone, it will likely just error?

The example that someone is then likely to bring up is a DSL, where a DSL is contained within some macro scope so that parsable but non-Julia expressions are how your interpretation of the AST is defined. However... that acts at compile-time. We are looking for a runtime definition since this is not a DSL or any type of language, but instead an IR to perform the manipulations in. Being a macro gives composibility issues (not to mention issues with things like Requires.jl), while being runtime values is something that can be used anywhere, such as inside of the macro's parsing functions.

But lastly, it doesn't even make sense to consider the AST here since the AST is not contextual, so there is not even a one-to-one mapping between Operation to AST. One has to choose how to lower it, which can be contextually dependent, since for example some packages may have dependencies be implicit in some places (like in a differential equation solve, `u` is actually `u(t)` but that needs to be interpreted as `u` the array... but not if you're acting on the solution of the differential equation where it's `u(t)`. Yes, computationally this difference makes sense, but you need a way to mathematically write down that expression and know how to automatically lower it if you're doing things like transforming to cascading ODE solves for example). Variables can have different context as well: there's a difference between independent variables and parameters when you take a derivative with respect to independent variables, since for parameters you don't necessarily know if they are functions of the independent variable or not (unless you have the context of course). The chain rule will be incorrect if this is not taken into account, and it will be very hard to things like simplifications on delay differential equations without the full representation. The Julia AST is essentially context flattened, which is great for a program, but ModelingToolkit IR doesn't represent programs.

>For example you can infer from the equations. ( if D(x) = Expr(...) then x must be a state variable.)

No, you can't. Using the derivative of a time-varying parameter is perfectly fine, and you may be writing a DAE there.

Hopefully that's enough to mull on. I can see how you can look at it and get an expression graph and think it's just Expr, but hopefully the example of computing a symbolic Jacobian of a user's numerical function demonstrates that this is a wildly different object.
