### 2019-10-03 JuliaDiffEq/ModelingToolkit.jl#186 @jamesjscully
Wow, thanks, this is great! It's much clearer now. So the benefits are:
1) adding Expr + Expr requires piracy. Operation avoids this
2) foo(x::Operation) will work for any Operation, whereas foo(x::Expr) will not work for every Expr
4) context is required to get the right differentiation behavior, and this context is stored in the context of Operation trees

It seems like you could use a type wrapper around an Expr with non-julia heads, ie Expr(:operation, op, args). That could enforce and preserve structure and avoid piracy, but I'm not sure how that would be materially different from using nested structs, and perhaps it is abusive to use program AST this way.
