### 2019-10-03 JuliaDiffEq/ModelingToolkit.jl#186 @jamesjscully
As far as the original issue: linking to the tutorial Chris mentioned at the beginning of README would definitely help with communicating the benefits of ModelingToolkit.jl
