### 2019-10-05 cd1d858
use simplified_expr for conversion
