+++ src/operations.jl
@@ -65,6 +65,5 @@ Base.convert(::Type{Operation}, x::Expression) = Operation(identity, Expression[
 Operation(x) = convert(Operation, x)
 
 #convert to Expr
-Base.Expr(op::Operation) = 
-    op.op isa Variable ? op.op.name : Expr(:call, Symbol(op.op), Expr.(op.args)...)
+Base.Expr(op::Operation) = simplified_expr(op)
 Base.convert(::Type{Expr},x::Operation) = Expr(x)
