### 2019-10-28 #52 @jpfairbanks
@harrisongrodin,
  Sorry for the delay in posting my notes.

I read the paper and think it should be accepted with minor revisions. I agree with most of @clagms comments.

Some broad considerations.

1. Why do scientific programmers use DSLs? I would say something about the difficulty and time consuming nature of correctly implementing scientific models as the complexity of the system grows.
2. Why not use the `Expr` type directly? As a user of MT, I can't really articulate it beyond "you want to use a different type because the usage is different".
3. You should discuss converting `Operation` to and from `Expr`, this is a common question for new users of MT and should be authoritatively presented in this paper.
4. What about sorted algebraic systems? Julia leverages type systems for scientific computing, shouldn't MT leverage sorts (the algebraic version on types)?
