+++ Project.toml
@@ -21,6 +21,14 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 julia = "1"
 Latexify = "0.11, 2"
 GeneralizedGenerated = "0.1.4"
+DiffEqBase = "6"
+DiffRules = "0.1"
+DocStringExtensions = "0.7, 0.8"
+Latexify = "0.12"
+MacroTools = "0.5"
+NaNMath = "0.3"
+SpecialFunctions = "0.7, 0.8"
+StaticArrays = "0.10, 0.11, 0.12"
 
 [extras]
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
