+++ Project.toml
@@ -19,12 +19,11 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
 julia = "1"
-Latexify = "0.11, 2"
-GeneralizedGenerated = "0.1.4"
 DiffEqBase = "6"
 DiffRules = "0.1"
 DocStringExtensions = "0.7, 0.8"
-Latexify = "0.12"
+Latexify = "0.11, 0.12"
+GeneralizedGenerated = "0.1.4"
 MacroTools = "0.5"
 NaNMath = "0.3"
 SpecialFunctions = "0.7, 0.8"
