### 2019-11-20 6a502aa
CompatHelper: bump compat for "GeneralizedGenerated" to "0.2"
