### 2019-12-03 JuliaDiffEq/ModelingToolkit.jl#184 @arnavs
I think elaborating on the `@eval eval(...)` would also be helpful here.
