### 2019-12-11 55b2bbf
new arg in nonlinear system function generation
