### 2019-12-11 1bebe9c
more generate_function controls
