+++ src/systems/nonlinear/nonlinear_system.jl
@@ -84,9 +84,9 @@ end
 (f::NLSysToExpr)(x) = convert(Expr, x)
 
 
-function generate_function(sys::NonlinearSystem, vs, ps, expression = Val{true}; version = nothing)
+function generate_function(sys::NonlinearSystem, vs, ps, expression = Val{true}; kwargs...)
     rhss = [eq.rhs for eq ∈ sys.eqs]
     vs′ = [clean(v) for v ∈ vs]
     ps′ = [clean(p) for p ∈ ps]
-    return build_function(rhss, vs′, ps′, (), NLSysToExpr(sys))
+    return build_function(rhss, vs′, ps′, (), NLSysToExpr(sys), expression; kwargs...)
 end
