### 2019-12-12 6288680
variable ordering
