### 2019-12-12 c89c417
Fix README example for https://github.com/JuliaDiffEq/ModelingToolkit.jl/issues/184
