### 2019-12-16 8e47afc
add support for gradient of a constant
