### 2019-12-16 47750c0
add a test for gradient dispatch for constants
