+++ test/derivatives.jl
@@ -60,3 +60,5 @@ jac = calculate_jacobian(sys)
 @test isequal(expand_derivatives(D(2t)), 2)
 @test isequal(expand_derivatives(D(2x)), 2D(x))
 @test isequal(expand_derivatives(D(x^2)), simplify_constants(2 * x * D(x)))
+
+@test isequal(ModelingToolkit.gradient(ModelingToolkit.constant(5), [1., 2., 3.,]), repeat(ModelingToolkit.Constant(0), 3))
