### 2019-12-20 17b2330
speed up MTK OOP vector usage
