### 2019-12-20 5bb1631
use map instead of broadcast and update README generated code
