### 2019-12-26 99293ff
Fix rational issues and Latexify tests
fixes https://github.com/JuliaDiffEq/ModelingToolkit.jl/issues/216
