+++ src/function_registration.jl
@@ -54,6 +54,7 @@ end
 
 # special cases
 Base.:^(x::Expression,y::T) where T <: Integer = Operation(Base.:^, Expression[x, y])
+Base.:^(x::Expression,y::T) where T <: Rational = Operation(Base.:^, Expression[x, y])
 
 @register Base.getindex(x,i)
 Base.getindex(x::Operation,i::Int64) = Operation(getindex,[x,i])
