+++ test/latexify.jl
@@ -3,7 +3,7 @@ using Latexify
 using ModelingToolkit
 
 ### Tips for generating latex tests:
-### Latexify has an unexported macro: 
+### Latexify has an unexported macro:
 ###
 ### Latexify.@generate_test
 ###
@@ -15,7 +15,7 @@ using ModelingToolkit
 ### This puts a ready-made test in your clipboard which you can paste into the
 ### test file.
 ###
-### Just be sure to remove all such macros before you commit a change since it 
+### Just be sure to remove all such macros before you commit a change since it
 ### will cause issues with Travis.
 
 @parameters t σ ρ β
@@ -26,12 +26,10 @@ eqs = [D(x) ~ σ*(y-x)*D(x-y)/D(z),
        0 ~ σ*x*(ρ-z)/10-y,
        D(z) ~ x*y - β*z]
 
-
-@test latexify(eqs) == 
+@test latexify(eqs) ==
 raw"\begin{align}
-\frac{dx(t)}{dt} =& \frac{\sigma \cdot \left( \mathrm{y}\left( t \right) - \mathrm{x}\left( t \right) \right) \cdot \frac{d\left(\mathrm{x}\left( t \right) - \mathrm{y}\left( t \right)\right)}{dt}}{\frac{dz(t)}{dt}} \\
+\mathrm{derivative}\left( \mathrm{x}\left( t \right), t \right) =& \frac{\sigma \cdot \left( \mathrm{y}\left( t \right) - \mathrm{x}\left( t \right) \right) \cdot \mathrm{derivative}\left( \mathrm{x}\left( t \right) - \mathrm{y}\left( t \right), t \right)}{\mathrm{derivative}\left( \mathrm{z}\left( t \right), t \right)} \\
 0 =& \frac{\sigma \cdot \mathrm{x}\left( t \right) \cdot \left( \rho - \mathrm{z}\left( t \right) \right)}{10} - \mathrm{y}\left( t \right) \\
-\frac{dz(t)}{dt} =& \mathrm{x}\left( t \right) \cdot \mathrm{y}\left( t \right) - \beta \cdot \mathrm{z}\left( t \right)
+\mathrm{derivative}\left( \mathrm{z}\left( t \right), t \right) =& \mathrm{x}\left( t \right) \cdot \mathrm{y}\left( t \right) - \beta \cdot \mathrm{z}\left( t \right)
 \end{align}
 "
-
