### 2019-12-30 9c92817
add a few missing methods for a patch
