### 2019-12-30 ae8e5cd
Link to tutorial in the README
Fixes https://github.com/JuliaDiffEq/ModelingToolkit.jl/issues/215
