+++ README.md
@@ -14,6 +14,11 @@ nicely with user-defined routines. Together, this is an abstract form of a
 scientific model that is easy for humans to generate but also easy for programs
 to manipulate.
 
+## Tutorial
+
+For an introductory tutorial to using ModelingToolkit.jl, please checkout
+[ModelingToolkit.jl, An IR and Compiler for Scientific Models](file:///C:/Users/accou/.julia/dev/DiffEqTutorials/html/ode_extras/01-ModelingToolkit.html).
+
 ## Introduction by Examples
 
 ### Example: ODE
