### 2020-01-13 280d6f2
Add muladd as a special case
