+++ src/function_registration.jl
@@ -55,6 +55,7 @@ end
 # special cases
 Base.:^(x::Expression,y::T) where T <: Integer = Operation(Base.:^, Expression[x, y])
 Base.:^(x::Expression,y::T) where T <: Rational = Operation(Base.:^, Expression[x, y])
+Base.muladd(a::Expression, b::Expression, c::Expression) = a * b + c
 
 @register Base.getindex(x,i)
 Base.getindex(x::Operation,i::Int64) = Operation(getindex,[x,i])
