### 2020-01-13 10112cf
Special case conj
Co-authored-by: "Shashi Gowda" <gowda@mit.edu>
Co-authored-by: "Yingbo Ma" <mayingbo5@gmail.com>
