### 2020-01-14 a3b6d13
Add promotion rule
