### 2020-02-06 8a65a34
fix array fast path on matrices
