+++ test/direct.jl
@@ -29,6 +29,21 @@ end
 @test all(isequal.(ModelingToolkit.gradient(eqs[1],[x,y,z]),[σ * -1,σ,0]))
 @test all(isequal.(ModelingToolkit.hessian(eqs[1],[x,y,z]),0))
 
+Joop,Jiip = eval.(ModelingToolkit.build_function(∂,[x,y,z],[σ,ρ,β],[t.op.name]))
+J = Joop([1.0,2.0,3.0],[1.0,2.0,3.0],1.0)
+@test J isa Matrix
+J2 = copy(J)
+Jiip(J2,[1.0,2.0,3.0],[1.0,2.0,3.0],1.0)
+@test J2 == J
+
+∂3 = cat(∂,∂,dims=3)
+Joop,Jiip = eval.(ModelingToolkit.build_function(∂3,[x,y,z],[σ,ρ,β],[t.op.name]))
+J = Joop([1.0,2.0,3.0],[1.0,2.0,3.0],1.0)
+@test size(J) == (3,3,2)
+J2 = copy(J)
+Jiip(J2,[1.0,2.0,3.0],[1.0,2.0,3.0],1.0)
+@test J2 == J
+
 # Function building
 
 @parameters σ() ρ() β()
