### 2020-02-06 fbcbef4
optimize sparse matrix formation
