+++ Project.toml
@@ -27,7 +27,7 @@ MacroTools = "0.5"
 NaNMath = "0.3"
 SpecialFunctions = "0.7, 0.8, 0.9, 0.10"
 StaticArrays = "0.10, 0.11, 0.12"
-julia = "1"
+julia = "1.2"
 
 [extras]
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
