### 2020-02-07 1687b58
fix non-square build_function
Fixes https://github.com/JuliaDiffEq/ModelingToolkit.jl/issues/235
