### 2020-02-07 98bf0dc
Add n-ary *, + rules
