### 2020-02-08 7835c1c
Add diff rule on constants
