### 2020-02-08 1423a4a
Fix expand_derivatives
