### 2020-02-08 0b3b4f5
handle the static array cases in build_function for non-vector outputs
