### 2020-02-08 f29ec53
better default constructors in the weird cases
