### 2020-02-17 d35ece2
CompatHelper: bump compat for "Latexify" to "0.13"
