+++ Project.toml
@@ -22,7 +22,7 @@ DiffEqBase = "6"
 DiffRules = "0.1, 1.0"
 DocStringExtensions = "0.7, 0.8"
 GeneralizedGenerated = "0.1.4, 0.2"
-Latexify = "0.11, 0.12"
+Latexify = "0.11, 0.12, 0.13"
 MacroTools = "0.5"
 NaNMath = "0.3"
 SpecialFunctions = "0.7, 0.8, 0.9, 0.10"
