+++ src/latexify_recipes.jl
@@ -2,7 +2,7 @@
     # Set default option values.
     env --> :align
 
-    # Convert both the left and right hand side to expressions of basic types 
+    # Convert both the left and right hand side to expressions of basic types
     # that latexify can deal with.
 
     rhs = getfield.(eqs, :rhs)
@@ -20,4 +20,28 @@
     lhs = [postwalk(x -> x isa Expr && x.args[1] == :Differential ? "\\frac{d\\left($(Latexify.latexraw(x.args[2]))\\right)}{d$iv}" : x, eq) for eq in lhs]
 
     return lhs, rhs
+end
+
+@latexrecipe function f(eqs::Vector{ModelingToolkit.DiffEq}; iv=:t)
+    # Set default option values.
+    env --> :align
+
+    # Convert both the left and right hand side to expressions of basic types
+    # that latexify can deal with.
+
+    rhs = getfield.(eqs, :rhs)
+    rhs = convert.(Expr, rhs)
+    rhs = [postwalk(x -> x isa ModelingToolkit.Constant ? x.value : x, eq) for eq in rhs]
+    rhs = [postwalk(x -> x isa Expr && length(x.args) == 1 ? x.args[1] : x, eq) for eq in rhs]
+    rhs = [postwalk(x -> x isa Expr && x.args[1] == :Differential && length(x.args[2].args) == 2 ? :($(Symbol(:d, x.args[2]))/($(Symbol(:d, x.args[2].args[2])))) : x, eq) for eq in rhs]
+    rhs = [postwalk(x -> x isa Expr && x.args[1] == :Differential ? "\\frac{d\\left($(Latexify.latexraw(x.args[2]))\\right)}{d$iv}" : x, eq) for eq in rhs]
+
+    var = getfield.(getfield.(eqs, :x),:name)
+    ns  = getfield.(eqs, :n)
+    lhs = [ns[i] == 1 ? "\\frac{d$(Latexify.latexraw(var[i]))}{d$iv}" : "\\frac{d^{$(ns[i])}$(Latexify.latexraw(var[i]))}{d$iv^{$(ns[i])}}" for i in 1:length(var)]
+    return lhs, rhs
+end
+
+@latexrecipe function f(sys::ModelingToolkit.ODESystem; iv=:t)
+    latexify(sys.eqs)
 end
