### 2020-02-20 dfb7ea3
some Latexify fixes
