+++ src/ModelingToolkit.jl
@@ -16,6 +16,7 @@ using MacroTools
 import MacroTools: splitdef, combinedef, postwalk, striplines
 import GeneralizedGenerated
 using DocStringExtensions
+using Base: RefValue
 
 """
 $(TYPEDEF)
@@ -91,7 +92,7 @@ include("simplify.jl")
 include("utils.jl")
 include("direct.jl")
 include("domains.jl")
-include("systems/diffeqs/diffeqsystem.jl")
+include("systems/diffeqs/odesystem.jl")
 include("systems/diffeqs/first_order_transform.jl")
 include("systems/nonlinear/nonlinear_system.jl")
 include("systems/pde/pdesystem.jl")
