### 2020-03-17 26fc511
Add diff rule for `one`
