### 2020-03-21 15d3c55
add get_variables
