+++ src/utils.jl
@@ -144,3 +144,15 @@ function vars!(vars, O)
 
     return vars
 end
+
+
+is_singleton(e) = hasproperty(e, :name) || hasproperty(e, :op) && hasproperty(e.op, :name)
+function get_variables(e::Expression, vars = nothing)
+  vars = isnothing(vars) ? [] : vars
+  if is_singleton(e)
+    push!(vars, e)
+  elseif hasproperty(e, :args) # because it can sometimes be 0
+    foreach(x -> get_variables(x, vars), e.args)
+  end
+  return vars
+end
