### 2020-03-21 92a992e
add in substitution logic
