### 2020-03-21 f4a94d2
Update src/utils.jl
Co-Authored-By: Christopher Rackauckas <accounts@chrisrackauckas.com>
