### 2020-03-21 1e2d526
Update src/utils.jl
Co-Authored-By: Christopher Rackauckas <accounts@chrisrackauckas.com>
