### 2020-03-21 4b3cc50
add generic method for is_singleton [ci skip]
