### 2020-03-21 ff8de3f
clean up get_variables [ci skip]
