### 2020-03-21 7cbf699
Update src/utils.jl
Co-Authored-By: Christopher Rackauckas <accounts@chrisrackauckas.com>
