### 2020-03-21 8c274a7
add get_variables
didnt commit uniqueness

add in substitution logic

add subs

Update src/utils.jl

Co-Authored-By: Christopher Rackauckas <accounts@chrisrackauckas.com>
Update src/utils.jl

Co-Authored-By: Christopher Rackauckas <accounts@chrisrackauckas.com>
add generic method for is_singleton [ci skip]

clean up get_variables [ci skip]

clean up substitution

Update src/utils.jl

Co-Authored-By: Christopher Rackauckas <accounts@chrisrackauckas.com>
fix up get variable and substitution


fix up getting and substitution
