+++ src/build_function.jl
@@ -1,0 +1,164 @@
+abstract type BuildTargets end
+struct JuliaTarget <: BuildTargets end
+struct StanTarget <: BuildTargets end
+struct CTarget <: BuildTargets end
+struct MATLABTarget <: BuildTargets end
+
+function build_function(args...;target = JuliaTarget(),kwargs...)
+  _build_function(target,args...;kwargs...)
+end
+
+function _build_function(target::JuliaTarget, rhss, vs, ps = (), args = (),
+                         conv = simplified_expr, expression = Val{true};
+                         checkbounds = false, constructor=nothing,
+                         linenumbers = true)
+    _vs = map(x-> x isa Operation ? x.op : x, vs)
+    _ps = map(x-> x isa Operation ? x.op : x, ps)
+    var_pairs   = [(u.name, :(u[$i])) for (i, u) ∈ enumerate(_vs)]
+    param_pairs = [(p.name, :(p[$i])) for (i, p) ∈ enumerate(_ps)]
+    (ls, rs) = zip(var_pairs..., param_pairs...)
+
+    var_eqs = Expr(:(=), build_expr(:tuple, ls), build_expr(:tuple, rs))
+
+    fname = gensym(:ModelingToolkitFunction)
+
+    X = gensym(:MTIIPVar)
+    if rhss isa SparseMatrixCSC
+        ip_sys_exprs = [:($X.nzval[$i] = $(conv(rhs))) for (i, rhs) ∈ enumerate(rhss.nzval)]
+    else
+        ip_sys_exprs = [:($X[$i] = $(conv(rhs))) for (i, rhs) ∈ enumerate(rhss)]
+    end
+
+    ip_let_expr = Expr(:let, var_eqs, build_expr(:block, ip_sys_exprs))
+
+    tuple_sys_expr = build_expr(:tuple, [conv(rhs) for rhs ∈ rhss])
+
+    if rhss isa Matrix
+        arr_sys_expr = build_expr(:vcat, [build_expr(:row,[conv(rhs) for rhs ∈ rhss[i,:]]) for i in 1:size(rhss,1)])
+        # : x because ??? what to do in the general case?
+        _constructor = constructor === nothing ? :(u isa ModelingToolkit.StaticArrays.StaticArray ? ModelingToolkit.StaticArrays.SMatrix{$(size(rhss)...)} :  x->(out = similar(typeof(u),$(size(rhss)...)); out .= x)) : constructor
+    elseif typeof(rhss) <: Array && !(typeof(rhss) <: Vector)
+        vector_form = build_expr(:vect, [conv(rhs) for rhs ∈ rhss])
+        arr_sys_expr = :(reshape($vector_form,$(size(rhss)...)))
+        _constructor = constructor === nothing ? :(u isa ModelingToolkit.StaticArrays.StaticArray ? ModelingToolkit.StaticArrays.SArray{$(size(rhss)...)} :  x->(out = similar(typeof(u),$(size(rhss)...)); out .= x)) : constructor
+    elseif rhss isa SparseMatrixCSC
+        vector_form = build_expr(:vect, [conv(rhs) for rhs ∈ nonzeros(rhss)])
+        arr_sys_expr = :(SparseMatrixCSC{eltype(u),Int}($(size(rhss)...), $(rhss.colptr), $(rhss.rowval), $vector_form))
+        # Static and sparse? Probably not a combo that will actually be hit, but give a default anyways
+        _constructor = constructor === nothing ? :(u isa ModelingToolkit.StaticArrays.StaticArray ? ModelingToolkit.StaticArrays.SMatrix{$(size(rhss)...)} : x->x) : constructor
+    else # Vector
[…]
