+++ src/systems/diffeqs/diffeqsystem.jl
@@ -289,9 +289,12 @@ $(SIGNATURES)
 Generate `ODESystem`, dependent variables, and parameters from an `ODEProblem`.
 """
 function modelingtoolkitize(prob::DiffEqBase.ODEProblem)
+    prob.f isa DiffEqBase.AbstractParameterizedFunction &&
+                            return (prob.f.sys, prob.f.sys.dvs, prob.f.sys.ps)
     @parameters t
     vars = [Variable(:x, i)(t) for i in eachindex(prob.u0)]
-    params = [Variable(:α,i; known = true)() for i in eachindex(prob.p)]
+    params = prob.p isa DiffEqBase.NullParameters ? [] :
+             [Variable(:α,i; known = true)() for i in eachindex(prob.p)]
     @derivatives D'~t
 
     rhs = [D(var) for var in vars]
