### 2020-03-25 00f805f
overload det
Fixes https://github.com/JuliaDiffEq/ModelingToolkit.jl/issues/258
