+++ test/operation_overloads.jl
@@ -1,0 +1,5 @@
+using ModelingToolkit
+using LinearAlgebra
+@variables a,b,c,d
+X = [a b;c d]
+det(X)
