+++ test/runtests.jl
@@ -3,6 +3,7 @@ using ModelingToolkit, Test
 @testset "Parsing Test" begin include("variable_parsing.jl") end
 @testset "Differentiation Test" begin include("derivatives.jl") end
 @testset "Simplify Test" begin include("simplify.jl") end
+@testset "Operation Overloads Test" begin include("operations_overloads.jl") end
 @testset "Direct Usage Test" begin include("direct.jl") end
 @testset "System Construction Test" begin include("system_construction.jl") end
 @testset "Build Targets Test" begin include("build_targets.jl") end
