+++ test/operation_overloads.jl
@@ -3,3 +3,5 @@ using LinearAlgebra
 @variables a,b,c,d
 X = [a b;c d]
 det(X)
+lu(X)
+inv(X)
