### 2020-04-09 c2bc515
some refactoring
