+++ src/systems/diffeqs/odesystem.jl
@@ -50,7 +50,7 @@ eqs = [D(x) ~ σ*(y-x),
 de = ODESystem(eqs)
 ```
 """
-struct ODESystem <: AbstractSystem
+struct ODESystem <: AbstractODESystem
     """The ODEs defining the system."""
     eqs::Vector{ODEExpr}
     """Independent variable."""
@@ -133,177 +133,3 @@ Base.:(==)(sys1::ODESystem, sys2::ODESystem) =
 independent_variables(sys::ODESystem) = Set{Variable}([sys.iv])
 dependent_variables(sys::ODESystem) = Set{Variable}(sys.dvs)
 parameters(sys::ODESystem) = Set{Variable}(sys.ps)
-
-function calculate_tgrad(sys::ODESystem)
-  isempty(sys.tgrad[]) || return sys.tgrad[]  # use cached tgrad, if possible
-  rhs = [detime_dvs(eq.rhs) for eq ∈ sys.eqs]
-  iv = sys.iv()
-  notime_tgrad = [expand_derivatives(ModelingToolkit.Differential(iv)(r)) for r in rhs]
-  tgrad = retime_dvs.(notime_tgrad,(sys.dvs,),iv)
-  sys.tgrad[] = tgrad
-  return tgrad
-end
-
-function calculate_jacobian(sys::ODESystem)
-    isempty(sys.jac[]) || return sys.jac[]  # use cached Jacobian, if possible
-    rhs = [eq.rhs for eq ∈ sys.eqs]
-
-    iv = sys.iv()
-    dvs = [dv(iv) for dv ∈ sys.dvs]
-
-    jac = expand_derivatives.(calculate_jacobian(rhs, dvs))
-    sys.jac[] = jac  # cache Jacobian
-    return jac
-end
-
-struct ODEToExpr
-    sys::ODESystem
-end
-function (f::ODEToExpr)(O::Operation)
-    if isa(O.op, Variable)
-        isequal(O.op, f.sys.iv) && return O.op.name  # independent variable
-        O.op ∈ f.sys.dvs        && return O.op.name  # dependent variables
-        isempty(O.args)         && return O.op.name  # 0-ary parameters
-        return build_expr(:call, Any[O.op.name; f.(O.args)])
-    end
-    return build_expr(:call, Any[Symbol(O.op); f.(O.args)])
-end
-(f::ODEToExpr)(x) = convert(Expr, x)
-
-function generate_tgrad(sys::ODESystem, dvs = sys.dvs, ps = sys.ps, expression = Val{true}; kwargs...)
-    tgrad = calculate_tgrad(sys)
-    return build_function(tgrad, dvs, ps, (sys.iv.name,), ODEToExpr(sys), expression; kwargs...)
-end
-
-function generate_jacobian(sys::ODESystem, dvs = sys.dvs, ps = sys.ps, expression = Val{true}; kwargs...)
-    jac = calculate_jacobian(sys)
-    return build_function(jac, dvs, ps, (sys.iv.name,), ODEToExpr(sys), expression; kwargs...)
-end
-
-function generate_function(sys::ODESystem, dvs = sys.dvs, ps = sys.ps, expression = Val{true}; kwargs...)
-    rhss = [deq.rhs for deq ∈ sys.eqs]
-    dvs′ = [clean(dv) for dv ∈ dvs]
-    ps′ = [clean(p) for p ∈ ps]
-    return build_function(rhss, dvs′, ps′, (sys.iv.name,), ODEToExpr(sys), expression; kwargs...)
-end
-
-function calculate_factorized_W(sys::ODESystem, simplify=true)
-    isempty(sys.Wfact[]) || return (sys.Wfact[],sys.Wfact_t[])
-
[…]
