### 2020-04-10 c3305d2
fix up SDESystem conversion
