### 2020-04-10 8584ab4
allow name for nonlinear system
