+++ src/systems/diffeqs/odesystem.jl
@@ -1,29 +1,3 @@
-isintermediate(eq::Equation) = !(isa(eq.lhs, Operation) && isa(eq.lhs.op, Differential))
-
-function flatten_differential(O::Operation)
-    @assert is_derivative(O) "invalid differential: $O"
-    is_derivative(O.args[1]) || return (O.args[1], O.op.x, 1)
-    (x, t, order) = flatten_differential(O.args[1])
-    isequal(t, O.op.x) || throw(ArgumentError("non-matching differentials on lhs: $t, $(O.op.x)"))
-    return (x, t, order + 1)
-end
-
-struct ODEExpr  # dⁿx/dtⁿ = rhs
-    x::Variable
-    n::Int
-    rhs::Expression
-end
-function Base.convert(::Type{ODEExpr},eq::Equation)
-    isintermediate(eq) && throw(ArgumentError("intermediate equation received"))
-    (x, t, n) = flatten_differential(eq.lhs)
-    (isa(t, Operation) && isa(t.op, Variable) && isempty(t.args)) ||
-        throw(ArgumentError("invalid independent variable $t"))
-    (isa(x, Operation) && isa(x.op, Variable) && length(x.args) == 1 && isequal(first(x.args), t)) ||
-        throw(ArgumentError("invalid dependent variable $x"))
-    return t.op, ODEExpr(x.op, n, eq.rhs)
-end
-Base.:(==)(a::ODEExpr, b::ODEExpr) = isequal((a.x, a.n, a.rhs), (b.x, b.n, b.rhs))
-
 """
 $(TYPEDEF)
 
@@ -50,7 +24,7 @@ de = ODESystem(eqs)
 """
 struct ODESystem <: AbstractODESystem
     """The ODEs defining the system."""
-    eqs::Vector{ODEExpr}
+    eqs::Vector{Equation}
     """Independent variable."""
     iv::Variable
     """Dependent (state) variables."""
@@ -83,55 +57,35 @@ struct ODESystem <: AbstractODESystem
     name::Symbol
 end
 
-function ODESystem(deqs::AbstractVector{ODEExpr}, iv, dvs, ps; name=gensym(:ODESystem))
+function ODESystem(deqs::AbstractVector{<:Equation}, iv, dvs, ps;
+                   name=gensym(:ODESystem))
+    iv′ = clean(iv)
+    dvs′ = [clean(dv) for dv ∈ dvs]
+    ps′ = [clean(p) for p ∈ ps]
     tgrad = RefValue(Vector{Expression}(undef, 0))
     jac = RefValue(Matrix{Expression}(undef, 0, 0))
     Wfact   = RefValue(Matrix{Expression}(undef, 0, 0))
     Wfact_t = RefValue(Matrix{Expression}(undef, 0, 0))
-    ODESystem(deqs, iv, dvs, ps, tgrad, jac, Wfact, Wfact_t, name)
+    ODESystem(deqs, iv′, dvs′, ps′, tgrad, jac, Wfact, Wfact_t, name)
 end
 
-function ODESystem(eqs; kwargs...)
-    reformatted = convert.(ODEExpr,eqs)
+var_from_nested_derivative(x) = var_from_nested_derivative(x,0)
+var_from_nested_derivative(x,i) = x.op isa Differential ? var_from_nested_derivative(x.args[1],i+1) : (x.op,i)
+iv_from_nested_derivative(x) = x.op isa Differential ? iv_from_nested_derivative(x.args[1]) : x.args[1].op
 
-    ivs = unique(r[1] for r ∈ reformatted)
+function ODESystem(eqs; kwargs...)
[…]

[…]
