+++ src/ModelingToolkit.jl
@@ -11,6 +11,8 @@ import GeneralizedGenerated
 using DocStringExtensions
 using Base: RefValue
 
+import TreeViews
+
 """
 $(TYPEDEF)
 
@@ -109,10 +111,9 @@ export simplify_constants
 export Operation, Expression
 export calculate_jacobian, generate_jacobian, generate_function
 export calculate_massmatrix, generate_diffusion_function
-export independent_variables, dependent_variables, parameters
+export independent_variable, states, parameters, equations
 export simplified_expr, eval_function
 export @register, @I
 export modelingtoolkitize
 export Variable, @variables, @parameters
-
 end # module

[…]
