### 2020-04-11 2c521bb
disable latexify tests
