### 2020-04-11 61f776d
Fix calculated functions on components
