+++ test/components.jl
@@ -18,7 +18,7 @@ lorenz2 = ODESystem(eqs,t,[x,y,z],[σ,ρ,β],name=:lorenz2)
 @variables a(t)
 connnectedeqs = [D(a) ~ a*states(lorenz1,:x)]
 
-connected = ODESystem(connnectedeqs,t,[a],[α],systems=[lorenz1,lorenz2],name=:connectedlorenz)
+connected1 = ODESystem(connnectedeqs,t,[a],[α],systems=[lorenz1,lorenz2],name=:connected1)
 
 @variables lorenz1′x(t) lorenz1′y(t) lorenz1′z(t) lorenz2′x(t) lorenz2′y(t) lorenz2′z(t)
 @parameters lorenz1′σ lorenz1′ρ lorenz1′β lorenz2′σ lorenz2′ρ lorenz2′β
@@ -31,6 +31,60 @@ eqs_flat = [D(a) ~ a*lorenz1′x,
             D(lorenz2′y) ~ lorenz2′x*(lorenz2′ρ-lorenz2′z)-lorenz2′y,
             0 ~ lorenz2′x + lorenz2′y + lorenz2′β*lorenz2′z]
 
-@test [x.name for x in states(connected)] == [:a,:lorenz1′x,:lorenz1′y,:lorenz1′z,:lorenz2′x,:lorenz2′y,:lorenz2′z]
-@test [x.name for x in parameters(connected)] == [:α,:lorenz1′σ,:lorenz1′ρ,:lorenz1′β,:lorenz2′σ,:lorenz2′ρ,:lorenz2′β]
-@test eqs_flat == equations(connected)
+@test [x.name for x in states(connected1)] == [:a,:lorenz1′x,:lorenz1′y,:lorenz1′z,:lorenz2′x,:lorenz2′y,:lorenz2′z]
+@test [x.name for x in parameters(connected1)] == [:α,:lorenz1′σ,:lorenz1′ρ,:lorenz1′β,:lorenz2′σ,:lorenz2′ρ,:lorenz2′β]
+@test eqs_flat == equations(connected1)
+
+connected2 = ODESystem(connnectedeqs,t,[a],[α],systems=[lorenz1,lorenz2],name=:connected2)
+
+@parameters γ
+@variables g(t)
+connnectedeqs2 = [D(g) ~ g*states(connected1,lorenz1,:x)]
+doublelevel = ODESystem(connnectedeqs2,t,[g],[γ],systems=[connected1,connected2],name=:doublelevel)
+
+@test [x.name for x in states(doublelevel)] == [:g,
+                                                :connected1′a,:connected1′lorenz1′x,:connected1′lorenz1′y,:connected1′lorenz1′z,:connected1′lorenz2′x,:connected1′lorenz2′y,:connected1′lorenz2′z,
+                                                :connected2′a,:connected2′lorenz1′x,:connected2′lorenz1′y,:connected2′lorenz1′z,:connected2′lorenz2′x,:connected2′lorenz2′y,:connected2′lorenz2′z]
+@test [x.name for x in parameters(doublelevel)] == [:γ,
+                                                   :connected1′α,:connected1′lorenz1′σ,:connected1′lorenz1′ρ,:connected1′lorenz1′β,:connected1′lorenz2′σ,:connected1′lorenz2′ρ,:connected1′lorenz2′β,
+                                                   :connected2′α,:connected2′lorenz1′σ,:connected2′lorenz1′ρ,:connected2′lorenz1′β,:connected2′lorenz2′σ,:connected2′lorenz2′ρ,:connected2′lorenz2′β]
+
[…]
