+++ src/systems/diffeqs/abstractodesystem.jl
@@ -1,19 +1,19 @@
 function calculate_tgrad(sys::AbstractODESystem)
   isempty(sys.tgrad[]) || return sys.tgrad[]  # use cached tgrad, if possible
-  rhs = [detime_dvs(eq.rhs) for eq ∈ sys.eqs]
+  rhs = [detime_dvs(eq.rhs) for eq ∈ equations(sys)]
   iv = sys.iv()
   notime_tgrad = [expand_derivatives(ModelingToolkit.Differential(iv)(r)) for r in rhs]
-  tgrad = retime_dvs.(notime_tgrad,(sys.dvs,),iv)
+  tgrad = retime_dvs.(notime_tgrad,(states(sys),),iv)
   sys.tgrad[] = tgrad
   return tgrad
 end
 
 function calculate_jacobian(sys::AbstractODESystem)
     isempty(sys.jac[]) || return sys.jac[]  # use cached Jacobian, if possible
-    rhs = [eq.rhs for eq ∈ sys.eqs]
+    rhs = [eq.rhs for eq ∈ equations(sys)]
 
     iv = sys.iv()
-    dvs = [dv(iv) for dv ∈ sys.dvs]
+    dvs = [dv(iv) for dv ∈ states(sys)]
 
     jac = expand_derivatives.(calculate_jacobian(rhs, dvs))
     sys.jac[] = jac  # cache Jacobian
@@ -34,18 +34,18 @@ function (f::ODEToExpr)(O::Operation)
 end
 (f::ODEToExpr)(x) = convert(Expr, x)
 
-function generate_tgrad(sys::AbstractODESystem, dvs = sys.dvs, ps = sys.ps, expression = Val{true}; kwargs...)
+function generate_tgrad(sys::AbstractODESystem, dvs = states(sys), ps = parameters(sys), expression = Val{true}; kwargs...)
     tgrad = calculate_tgrad(sys)
     return build_function(tgrad, dvs, ps, (sys.iv.name,), ODEToExpr(sys), expression; kwargs...)
 end
 
-function generate_jacobian(sys::AbstractODESystem, dvs = sys.dvs, ps = sys.ps, expression = Val{true}; kwargs...)
+function generate_jacobian(sys::AbstractODESystem, dvs = states(sys), ps = parameters(sys), expression = Val{true}; kwargs...)
     jac = calculate_jacobian(sys)
     return build_function(jac, dvs, ps, (sys.iv.name,), ODEToExpr(sys), expression; kwargs...)
 end
 
-function generate_function(sys::AbstractODESystem, dvs = sys.dvs, ps = sys.ps, expression = Val{true}; kwargs...)
-    rhss = [deq.rhs for deq ∈ sys.eqs]
+function generate_function(sys::AbstractODESystem, dvs = states(sys), ps = parameters(sys), expression = Val{true}; kwargs...)
+    rhss = [deq.rhs for deq ∈ equations(sys)]
     dvs′ = [clean(dv) for dv ∈ dvs]
     ps′ = [clean(p) for p ∈ ps]
     return build_function(rhss, dvs′, ps′, (sys.iv.name,), ODEToExpr(sys), expression; kwargs...)
@@ -75,7 +75,7 @@ function calculate_factorized_W(sys::AbstractODESystem, simplify=true)
     (Wfact,Wfact_t)
 end
 
-function generate_factorized_W(sys::AbstractODESystem, vs = sys.dvs, ps = sys.ps, simplify=true, expression = Val{true}; kwargs...)
+function generate_factorized_W(sys::AbstractODESystem, vs = states(sys), ps = parameters(sys), simplify=true, expression = Val{true}; kwargs...)
     (Wfact,Wfact_t) = calculate_factorized_W(sys,simplify)
     siz = size(Wfact)
     constructor = :(x -> begin
[…]
