### 2020-04-11 30a8533
simplify a bit first
