+++ test/variable_parsing.jl
@@ -17,9 +17,9 @@ z1 = Variable(:z)(t)
 end
 @parameters σ(..)
 
-t1 = Variable(:t; known = true)()
-s1 = Variable(:s; known = true)()
-σ1 = Variable(:σ; known = true)
+t1 = Variable(:t)()
+s1 = Variable(:s)()
+σ1 = Variable(:σ)
 @test isequal(t1, t)
 @test isequal(s1, s)
 @test isequal(σ1, σ)
@@ -42,12 +42,12 @@ convert(Expression, :($x == 0 ? $y : $x))
 end
 @parameters σ[1:2](..)
 
-t1 = [Variable(:t, 1; known = true)(),
-      Variable(:t, 2; known = true)()]
-s1 = [Variable(:s, 1, 1; known = true)() Variable(:s, 1, 2; known = true)()
-      Variable(:s, 3, 1; known = true)() Variable(:s, 3, 2; known = true)()]
-σ1 = [Variable(:σ, 1; known = true),
-      Variable(:σ, 2; known = true)]
+t1 = [Variable(:t, 1)(),
+      Variable(:t, 2)()]
+s1 = [Variable(:s, 1, 1)() Variable(:s, 1, 2)()
+      Variable(:s, 3, 1)() Variable(:s, 3, 2)()]
+σ1 = [Variable(:σ, 1),
+      Variable(:σ, 2)]
 @test isequal(t1, t)
 @test isequal(s1, s)
 @test isequal(σ1, σ)
