### 2020-04-11 f089bbd
make variables typed
