### 2020-04-11 ce79f03
add unit evaluation of equations
