### 2020-04-11 f18f3d2
Make sized variable parsing make actual sized variables
