### 2020-04-11 2b1c0eb
add dot syntax, variable searching in scope, and variable->operation
Fixes https://github.com/SciML/ModelingToolkit.jl/issues/130 and https://github.com/SciML/ModelingToolkit.jl/issues/276
