### 2020-04-11 205a264
refactor to AbstractSystem interface and create OptimizationSystem
