+++ test/odesystem.jl
@@ -137,7 +137,7 @@ lowered_eqs = [D(u_tt) ~ 2u_tt + u_t + x_t + 1
 @test de1 == ODESystem(lowered_eqs)
 
 # issue #219
-@test de1.dvs == [ModelingToolkit.var_from_nested_derivative(eq.lhs)[1] for eq in de1.eqs] == ODESystem(lowered_eqs).dvs
+@test de1.states == [ModelingToolkit.var_from_nested_derivative(eq.lhs)[1] for eq in de1.eqs] == ODESystem(lowered_eqs).states
 
 test_diffeq_inference("first-order transform", de1, t, [u_tt, x_t, u_t, u, x], [])
 du = zeros(5)
@@ -168,41 +168,6 @@ de = ODESystem(eqs)
     f = eval(generate_function(de, [x,y], [A,B,C])[1])
     du ≈ f([1.0,2.0], [1,2,3], 0.0)
 end
-
-# Now nonlinear system with only variables
-@variables x y z
-@parameters σ ρ β
-
-# Define a nonlinear system
-eqs = [0 ~ σ*(y-x),
-       0 ~ x*(ρ-z)-y,
-       0 ~ x*y - β*z]
-ns = NonlinearSystem(eqs, [x,y,z])
-jac = calculate_jacobian(ns)
-@testset "nlsys jacobian" begin
-    @test isequal(jac[1,1], σ * -1)
-    @test isequal(jac[1,2], σ)
-    @test isequal(jac[1,3], 0)
-    @test isequal(jac[2,1], ρ - z)
-    @test isequal(jac[2,2], -1)
-    @test isequal(jac[2,3], x * -1)
-    @test isequal(jac[3,1], y)
-    @test isequal(jac[3,2], x)
-    @test isequal(jac[3,3], -1 * β)
-end
-nlsys_func = generate_function(ns, [x,y,z], [σ,ρ,β])
-jac_func = generate_jacobian(ns)
-f = @eval eval(nlsys_func)
-
-# Intermediate calculations
-# Define a nonlinear system
-eqs = [0 ~ σ*a,
-       0 ~ x*(ρ-z)-y,
-       0 ~ x*y - β*z]
-ns = NonlinearSystem(eqs, [x,y,z])
-nlsys_func = generate_function(ns, [x,y,z], [σ,ρ,β])
-jac = calculate_jacobian(ns)
-jac = generate_jacobian(ns)
 
 function lotka(u,p,t)
   x = u[1]
