+++ test/optimizationsystem.jl
@@ -1,0 +1,19 @@
+using ModelingToolkit
+@variables x y
+@parameters a b
+loss = (a - x)^2 + b * (y - x^2)^2
+sys1 = OptimizationSystem(loss,[x,y],[a,b],name=:sys1)
+sys2 = OptimizationSystem(loss,[x,y],[a,b],name=:sys2)
+
+@variables z
+@parameters β
+loss2 = sys1.x - sys2.y + z*β
+combinedsys = OptimizationSystem(loss,[z],[β],systems=[sys1,sys2],name=:combinedsys)
+
+equations(combinedsys)
+states(combinedsys)
+parameters(combinedsys)
+
+calculate_hessian(combinedsys)
+generate_function(combinedsys)
+generate_hessian(combinedsys)
