### 2020-04-11 02ed45d
add and test automatic multithreading
