+++ test/runtests.jl
@@ -15,4 +15,5 @@ using ModelingToolkit, Test
 @testset "Constraints Test" begin include("constraints.jl") end
 @testset "PDE Construction Test" begin include("pde.jl") end
 @testset "Distributed Test" begin include("distributed.jl") end
+@testset "Test Big System Usage" begin include("bigsystem.jl") end
 #@testset "Latexify recipes Test" begin include("latexify.jl") end
