### 2020-04-11 3c88997
setup test for Travis
