+++ test/bigsystem.jl
@@ -1,4 +1,4 @@
-using ModelingToolkit, LinearAlgebra
+using ModelingToolkit, LinearAlgebra, SparseArrays
 
 # Define the constants for the PDE
 const α₂ = 1.0
@@ -45,6 +45,13 @@ end
 
 f(du,u,nothing,0.0)
 
-ModelingToolkit.build_function(du,u,multithread=true);
-jac = ModelingToolkit.jacobian(vec(du),vec(u),simplify=false)
-ModelingToolkit.build_function(vec(jac),u,multithread=true);
+multithreadedf = eval(ModelingToolkit.build_function(du,u,multithread=true)[2])
+_du = rand(N,N,3)
+_u = rand(N,N,3)
+multithreadedf(_du,_u)
+
+jac = sparse(ModelingToolkit.jacobian(vec(du),vec(u),simplify=false))
+multithreadedjac = eval(ModelingToolkit.build_function(vec(jac),u,multithread=true)[2])
+
+#_jac = similar(jac,Float64)
+#multithreadedjac(_jac,_u)
