+++ test/runtests.jl
@@ -1,19 +1,19 @@
-using ModelingToolkit, Test
+using SafeTestsets
 
-@testset "Parsing Test" begin include("variable_parsing.jl") end
-@testset "Differentiation Test" begin include("derivatives.jl") end
-@testset "Simplify Test" begin include("simplify.jl") end
-@testset "Operation Overloads Test" begin include("operation_overloads.jl") end
-@testset "Direct Usage Test" begin include("direct.jl") end
-@testset "ODESystem Test" begin include("odesystem.jl") end
-@testset "Mass Matrix Test" begin include("mass_matrix.jl") end
-@testset "SDESystem Test" begin include("sdesystem.jl") end
-@testset "NonlinearSystem Test" begin include("nonlinearsystem.jl") end
-@testset "OptimizationSystem Test" begin include("optimizationsystem.jl") end
-@testset "Build Targets Test" begin include("build_targets.jl") end
-@testset "Domain Test" begin include("domains.jl") end
-@testset "Constraints Test" begin include("constraints.jl") end
-@testset "PDE Construction Test" begin include("pde.jl") end
-@testset "Distributed Test" begin include("distributed.jl") end
-@testset "Test Big System Usage" begin include("bigsystem.jl") end
+@safetestset "Parsing Test" begin include("variable_parsing.jl") end
+@safetestset "Differentiation Test" begin include("derivatives.jl") end
+@safetestset "Simplify Test" begin include("simplify.jl") end
+@safetestset "Operation Overloads Test" begin include("operation_overloads.jl") end
+@safetestset "Direct Usage Test" begin include("direct.jl") end
+@safetestset "ODESystem Test" begin include("odesystem.jl") end
+@safetestset "Mass Matrix Test" begin include("mass_matrix.jl") end
+@safetestset "SDESystem Test" begin include("sdesystem.jl") end
+@safetestset "NonlinearSystem Test" begin include("nonlinearsystem.jl") end
+@safetestset "OptimizationSystem Test" begin include("optimizationsystem.jl") end
+@safetestset "Build Targets Test" begin include("build_targets.jl") end
+@safetestset "Domain Test" begin include("domains.jl") end
+@safetestset "Constraints Test" begin include("constraints.jl") end
+@safetestset "PDE Construction Test" begin include("pde.jl") end
+@safetestset "Distributed Test" begin include("distributed.jl") end
+@safetestset "Test Big System Usage" begin include("bigsystem.jl") end
 #@testset "Latexify recipes Test" begin include("latexify.jl") end
