### 2020-04-12 0d4437f
CompatHelper: add new compat entry for "SafeTestsets" at version "0.0.1"
