+++ Project.toml
@@ -28,6 +28,7 @@ GeneralizedGenerated = "0.1.4, 0.2"
 Latexify = "0.11, 0.12, 0.13"
 MacroTools = "0.5"
 NaNMath = "0.3"
+SafeTestsets = "0.0.1"
 SpecialFunctions = "0.7, 0.8, 0.9, 0.10"
 StaticArrays = "0.10, 0.11, 0.12"
 TreeViews = "0.3"
