### 2020-04-12 5cfa982
fix latexification
Fixes https://github.com/SciML/ModelingToolkit.jl/issues/279
