### 2020-04-12 8092448
setup Latexify tests using 0.13.0
