### 2020-04-12 3a34965
Setup ReactionSystem on AbstractSystem interface
