### 2020-04-12 7378b39
Generalize LaTeXify for AbstractSystem
