+++ src/ModelingToolkit.jl
@@ -4,6 +4,7 @@ using DiffEqBase, Distributed
 using StaticArrays, LinearAlgebra, SparseArrays
 using Latexify, Unitful
 using MacroTools
+using Parameters: @unpack
 
 using Base.Threads
 import MacroTools: splitdef, combinedef, postwalk, striplines
