+++ src/systems/reaction/reactionsystem.jl
@@ -1,57 +1,83 @@
-struct Reaction
+struct Reaction{T <: Number}
     rate
-    reactants::Vector{Operation}
+    substrates::Vector{Operation}
     products::Vector{Operation}
+    substoich::Vector{T}
+    prodstoich::Vector{T}
 end
 
 struct ReactionSystem <: AbstractSystem
     eqs::Vector{Reaction}
     iv::Variable
-    states::Vector{Variable}
-    ps::Vector{Variable}
+    species::Vector{Variable}
+    params::Vector{Variable}
     name::Symbol
     systems::Vector{ReactionSystem}
 end
 
-function ReactionSystem(eqs,iv,dvs,ps;
+function ReactionSystem(eqs, iv, species, params;
                         systems = ReactionSystem[],
-                        name=gensym(:ReactionSystem))
-    ReactionSystem(eqs,iv,convert.(Variable,dvs),convert.(Variable,ps),name,systems)
+                        name = gensym(:ReactionSystem))
+
+    ReactionSystem(eqs, iv, convert.(Variable,species), 
+                   convert.(Variable,params), name, systems)
 end
 
-# TODO: Make it do the combinatorics stuff
-reaction_expr(reactants) = *(reactants...)
+# Calculate the ODE rate law
+function oderatelaw(rx)
+    @unpack rate, substrates, substoich = rx    
+    rl   = rate
+    coef = one(eltype(substoich))
+    for (i,stoich) in enumerate(substoich)
+        coef *= factorial(stoich)        
+        rl   *= (stoich != one(stoich)) ? substrates[i]^stoich : substrates[i]
+    end
+    (coef != one(coef)) && (rl /= coef)
+
+    rl
+end
 
 function essemble_drift(rs)
-    D = Differential(rs.iv())
-    eqs = [D(x(rs.iv())) ~ 0 for x in rs.states]
+    D   = Differential(rs.iv())
+    eqs = [D(x(rs.iv())) ~ 0 for x in rs.species]
+    species_to_idx = Dict(enumerate(rs.species))
 
-    for rx in rs.eqs
-        for reactant in rx.reactants
-            i = findfirst(x->x == reactant.op,rs.states)
-            eqs[i] = Equation(eqs[i].lhs,eqs[i].rhs - rx.rate * reaction_expr(rx.reactants))
+    # note, this should really use the net stoichiometry to avoid 
+    # adding and substracting the same term for A + X -> B + X
+    for rx in rs.eqs        
+        rl     = oderatelaw(rx.rate, rx.substrates, rx.substoich)
+        stoich = rx.substoich
+        for (sidx,substrate) in enumerate(rx.substrates)
+            i      = species_to_idx[substrate.op]
+            eqs[i] = Equation(eqs[i].lhs, eqs[i].rhs - stoich[sidx]*rl)
         end
 
-        for product in rx.products
-            i = findfirst(x->x == product.op,rs.states)
-            eqs[i] = Equation(eqs[i].lhs,eqs[i].rhs + rx.rate * reaction_expr(rx.reactants))
+        stoich = rx.prodstoich
+        for (pidx,product) in enumerate(rx.products)
+            i      = species_to_idx[product.op]
+            eqs[i] = Equation(eqs[i].lhs,eqs[i].rhs + stoich[pidx]*rl)
         end
     end
+
     eqs
 end
 
 function essemble_diffusion(rs)
-    eqs = Expression[Constant(0) for x in rs.states, y in rs.eqs]
+    eqs = Expression[Constant(0) for x in rs.species, y in rs.eqs]
[…]
