+++ src/systems/nonlinear/nonlinearsystem.jl
@@ -71,3 +71,16 @@ function generate_function(sys::NonlinearSystem, vs = states(sys), ps = paramete
     ps′ = convert.(Variable,ps)
     return build_function(rhss, vs′, ps′, (), NLSysToExpr(sys), expression; kwargs...)
 end
+
+function DiffEqBase.NonlinearProblem{iip}(sys::NonlinearSystem,u0map,tspan,
+                                          parammap=DiffEqBase.NullParameters();
+                                          version = nothing, jac = false,
+                                          kwargs...) where iip
+    dvs = states(sys)
+    ps = parameters(sys)
+
+    f = generate_function(sys)
+    u0 = varmap_to_vars(u0map,dvs)
+    p = varmap_to_vars(parammap,ps)
+    NonlinearProblem(f,u0,tspan,p;kwargs...)
+end
