### 2020-04-13 0e25e2c
Expose function generation keyword arguments from problem interface
