+++ src/systems/optimization/optimizationsystem.jl
@@ -70,11 +70,14 @@ function DiffEqBase.OptimizationProblem{iip}(sys::OptimizationSystem,
                                           parammap=DiffEqBase.NullParameters();
                                           u0=nothing, lb=nothing, ub=nothing,
                                           hes = false,
+                                          checkbounds = false,
+                                          linenumbers = true, multithread=false,
                                           kwargs...) where iip
     dvs = states(sys)
     ps = parameters(sys)
 
-    f = generate_function(sys)
+    f = generate_function(sys,checkbounds=checkbounds,linenumbers=linenumbers,
+                              multithread=multithread)
     u0 = varmap_to_vars(u0,dvs)
     p = varmap_to_vars(parammap,ps)
     lb = varmap_to_vars(lb,dvs)
