+++ src/systems/nonlinear/nonlinearsystem.jl
@@ -74,11 +74,15 @@ end
 
 function DiffEqBase.NonlinearProblem{iip}(sys::NonlinearSystem,u0map,tspan,
                                           parammap=DiffEqBase.NullParameters();
-                                          jac = false,kwargs...) where iip
+                                          jac = false,
+                                          checkbounds = false,
+                                          linenumbers = true, multithread=false,
+                                          kwargs...) where iip
     dvs = states(sys)
     ps = parameters(sys)
 
-    f = generate_function(sys)
+    f = generate_function(sys;checkbounds=checkbounds,linenumbers=linenumbers,
+                              multithread=multithread)
     u0 = varmap_to_vars(u0map,dvs)
     p = varmap_to_vars(parammap,ps)
     NonlinearProblem(f,u0,tspan,p;kwargs...)
