### 2020-04-13 8d2e038
Expose sparse matrix generation in higher level prob constructors
