### 2020-04-13 5cb6d35
fix rates and minimize generated expression ops
