+++ src/systems/reaction/reactionsystem.jl
@@ -1,26 +1,57 @@
-struct Reaction{T <: Number}
+struct Reaction{S <: Variable, T <: Number}
     rate
     substrates::Vector{Operation}
     products::Vector{Operation}
     substoich::Vector{T}
-    prodstoich::Vector{T}
+    prodstoich::Vector{T}    
+    netstoich::Vector{Pair{S,T}}
+    only_use_rate::Bool
 end
+
+function Reaction(rate, subs, prods, substoich, prodstoich; 
+                  netstoich=nothing, only_use_rate=false, kwargs...)
+
+    ns = isnothing(netstoich) ? get_netstoich(subs, prods, substoich, prodstoich) : netstoich
+    Reaction(rate, subs, prods, substoich, prodstoich, ns, only_use_rate)
+end
+
+# three argument constructor assumes stoichiometric coefs are one and integers
+Reaction(rate, subs, prods; kwargs...) = Reaction(rate, subs, prods,
+                                                  ones(Int,length(subs)),
+                                                  ones(Int,length(prods)); 
+                                                  kwargs...)
+
+# calculates the net stoichiometry of a reaction as a vector of pairs (sub,substoich)
+function get_netstoich(subs, prods, sstoich, pstoich)
+    # stoichiometry as a Dictionary
+    nsdict = Dict(sub.op => -sstoich[i] for (i,sub) in enumerate(subs))
+    for (i,p) in enumerate(prods)
+        coef = pstoich[i]
+        prod = p.op
+        @inbounds nsdict[prod] = haskey(nsdict, prod) ? nsdict[prod] + coef : coef
+    end
+
+    # stoichiometry as a vector
+    ns = [el for el in nsdict if el[2] != zero(el[2])]
+
+    ns
+end
+
 
 struct ReactionSystem <: AbstractSystem
     eqs::Vector{Reaction}
     iv::Variable
-    species::Vector{Variable}
-    params::Vector{Variable}
+    states::Vector{Variable}
+    ps::Vector{Variable}
     name::Symbol
     systems::Vector{ReactionSystem}
 end
 
-function ReactionSystem(eqs, iv, species, params;
-                        systems = ReactionSystem[],
-                        name = gensym(:ReactionSystem))
+function ReactionSystem(eqs, iv, species, params; systems = ReactionSystem[],
+                                                  name = gensym(:ReactionSystem))
 
-    ReactionSystem(eqs, iv, convert.(Variable,species), 
-                   convert.(Variable,params), name, systems)
+    ReactionSystem(eqs, iv, convert.(Variable,species), convert.(Variable,params), 
+                   name, systems)
 end
 
 # Calculate the ODE rate law
@@ -37,61 +68,54 @@ function oderatelaw(rx)
     rl
 end
 
-function essemble_drift(rs)
+function assemble_drift(rs)
     D   = Differential(rs.iv())
-    eqs = [D(x(rs.iv())) ~ 0 for x in rs.species]
-    species_to_idx = Dict(enumerate(rs.species))
+    eqs = [D(x(rs.iv())) ~ 0 for x in rs.states]
+    species_to_idx = Dict((x => i for (i,x) in enumerate(rs.states)))
 
-    # note, this should really use the net stoichiometry to avoid 
-    # adding and substracting the same term for A + X -> B + X
     for rx in rs.eqs        
-        rl     = oderatelaw(rx.rate, rx.substrates, rx.substoich)
-        stoich = rx.substoich
-        for (sidx,substrate) in enumerate(rx.substrates)
[…]
