### 2020-04-13 03478b1
support turning off mass action terms in ratelaw
