### 2020-04-13 d38aaeb
CompatHelper: add new compat entry for "ArrayInterface" at version "2.8"
