+++ docs/src/tracing.md
@@ -1,0 +1,5 @@
+# Tracing and Modelingtoolkitization
+
+## Direct Tracing
+
+## ModelingToolkitize
