+++ docs/src/highlevel.md
@@ -51,19 +51,41 @@ p  = [σ => 10.0
       ρ => 28.0
       β => 8/3]
 tspan = (0.0,100.0)
-prob = ODEProblem(sys,u0,tspan,p;jac=true)
+prob = ODEProblem(sys,u0,tspan,p;jac=true,sparse=true)
 ```
 
 Note that the additional `jac=true` tells the system to symbolically generate
-an optimized Jacobian function to enhance the differential equation solvers.
+an optimized Jacobian function to enhance the differential equation solvers,
+and `sparse` tells it to build the ODEProblem with all of the enhancements
+setup for sparse Jacobians.
 
-### Example 2: Building a Component-Based ODEProblem with Sparse Jacobians
+### Example 2: Building a Component-Based ODEProblem
+
+In addition, we can then use ModelingToolkit to compose multiple ODE subsystems.
+Let's define two interacting Lorenz equations:
+
+```julia
+lorenz1 = ODESystem(eqs,name=:lorenz1)
+lorenz2 = ODESystem(eqs,name=:lorenz2)
+
+@variables α
+@parameters γ
+connections = [0 ~ lorenz1.x + lorenz2.y + sin(α*γ)]
+connected = ODESystem(connections,[α],[γ],systems=[lorenz1,lorenz2])
+```
+
+which is now a differential-algebraic equation (DAE) of 7 variables which has
+two independent Lorenz systems and an algebraic equation that determines `α`
+such that an implicit constraint holds. We can then define the resulting
+`ODEProblem` and send it over to DifferentialEquations.jl.
 
 ### Example 3: Building Nonlinear Systems to Solve with NLsolve.jl
 
-We can also build nonlinear systems. Let's say we wanted to solve for the steady
-state of the previous ODE. This is the nonlinear system defined by where the
-derivatives are zero. We use (unknown) variables for our nonlinear system.
+In this example we will go one step deeper and showcase the direct function
+generation capabilities in ModelingToolkit.jl to build nonlinear systems.
+Let's say we wanted to solve for the steady state of the previous ODE. This is
+the nonlinear system defined by where the derivatives are zero. We use (unknown)
+variables for our nonlinear system.
 
 ```julia
 using ModelingToolkit
@@ -150,6 +172,7 @@ which gives:
 Now we can call `nlsolve` by enclosing our parameters into the functions:
 
 ```julia
+using NLsolve
 nlsolve((out, x) -> f(out, x, params), (out, x) -> j!(out, x, params), ones(3))
 ```
 
@@ -174,7 +197,7 @@ Base.:~(::Expression, ::Expression)
 
 ## Additional High Level Explanations and Tips
 
-## The Auto-Detecting System Constructors
+### The Auto-Detecting System Constructors
 
 For the high level interface, the system constructors such as `ODESystem` have
 high level constructors which just take in the required equations and automatically
@@ -184,6 +207,114 @@ The following high level constructors exist:
 ```julia
 ODESystem(eqs)
 NonlinearSystem(eqs)
+```
+
+### Direct Tracing
+
+Because the ModelingToolkit `Expression` types obey Julia-semantics, one can
+directly transform existing Julia functions into ModelingToolkit symbolic
[…]
