+++ docs/src/tracing.md
@@ -1,5 +1,0 @@
-# Tracing and Modelingtoolkitization
-
-## Direct Tracing
-
-## ModelingToolkitize
