### 2020-04-13 e2b9cd3
major docs cleanup
