+++ src/systems/diffeqs/first_order_transform.jl
@@ -13,7 +13,7 @@ function flatten_differential(O::Operation)
 end
 
 """
-ode_order_lowering(sys::ODESystem)
+$(TYPEDSIGNATURES)
 
 Takes a Nth order ODESystem and returns a new ODESystem written in first order
 form by defining new variables which represent the N-1 derivatives.
