### 2020-04-13 f066cb6
fix Wfact italics
