+++ src/systems/diffeqs/sdesystem.jl
@@ -53,7 +53,7 @@ struct SDESystem <: AbstractODESystem
     """
     Wfact::RefValue{Matrix{Expression}}
     """
-    Wfact\_t matrix. Note: this field will not be defined until
+    Wfact_t matrix. Note: this field will not be defined until
     [`generate_factorized_W`](@ref) is called on the system.
     """
     Wfact_t::RefValue{Matrix{Expression}}
