### 2020-04-13 02df4d1
add lowering integration test
