### 2020-04-13 77223a7
fix composed expression generation
