### 2020-04-13 8e29750
fix docs typos
