### 2020-04-13 4be15f7
little docs improvements
