### 2020-04-13 4268cd3
cleanup equality with one checks
