### 2020-04-13 9e7d373
shorten conditional logic
