### 2020-04-13 e9ee923
handle multiplication by neg one
