### 2020-04-13 bb93049
Generalize build_function
No longer requires u,p and treats those two arguments differently. Now it's just a list of arguments, which can understand if a variable is scalar or not.

Fixes https://github.com/SciML/ModelingToolkit.jl/issues/260
