### 2020-04-13 fc03508
fix build function test
