### 2020-04-13 98d5720
allow arrays of arrays of stuff in build_function iip
Fixes https://github.com/SciML/ModelingToolkit.jl/issues/262
