+++ test/build_function_arrayofarray.jl
@@ -1,0 +1,209 @@
+using ModelingToolkit, Test, SparseArrays
+@variables a b c
+
+# Auxiliary Functions and Constants
+get_sparsity_pattern(h::Array{Expression}) = sparse(Int64.(map(~, h .=== ModelingToolkit.Constant(0))))
+get_sparsity_pattern(h::SparseMatrixCSC{Expression,Int64}) = sparse(Int64.(map(~, h .=== ModelingToolkit.Constant(0))))
+get_sparsity_pattern(h::SparseVector{Expression,Int64}) = sparse(Int64.(map(~, h .=== ModelingToolkit.Constant(0))))
+
+input = [1, 2, 3]
+
+# ===== Dense tests =====
+# Arrays of Matrices
+h_dense_arraymat = [[a 1; b 0], [0 0; 0 0], [a c; 1 0]] # empty array support required
+function h_dense_arraymat_julia(x)
+    a, b, c = x
+    return [[a[1] 1; b[1] 0], [0 0; 0 0], [a[1] c[1]; 1 0]]
+end
+function h_dense_arraymat_julia!(out, x)
+    a, b, c = x
+    out[1] .= [a[1] 1; b[1] 0]
+    out[2] .= [0 0; 0 0]
+    out[3] .= [a[1] c[1]; 1 0]
+end
+
+h_dense_arraymat_str = ModelingToolkit.build_function(h_dense_arraymat, [a, b, c])
+h_dense_arraymat_oop = eval(h_dense_arraymat_str[1])
+h_dense_arraymat_ip! = eval(h_dense_arraymat_str[2])
+out_1_arraymat = [Array{Int64}(undef, 2, 2) for i in 1:3]
+out_2_arraymat = [similar(x) for x in out_1_arraymat]
+julia_dense_arraymat = h_dense_arraymat_julia(input)
+mtk_dense_arraymat = h_dense_arraymat_oop(input)
+@test_broken julia_dense_arraymat == mtk_dense_arraymat
+h_dense_arraymat_julia!(out_1_arraymat, input)
+h_dense_arraymat_ip!(out_2_arraymat, input)
+@test out_1_arraymat == out_2_arraymat
+
+# Arrays of 1D Vectors
+h_dense_arrayvec = [[a, 0, c], [0, 0, 0], [1, a, b]] # same for empty vectors, etc.
+function h_dense_arrayvec_julia(x)
+    a, b, c = x
+    return [[a[1], 0, c[1]], [0, 0, 0], [1, a[1], b[1]]]
+end
+function h_dense_arrayvec_julia!(out, x)
+    a, b, c = x
+    out[1] .= [a[1], 0, c[1]]
+    out[2] .= [0, 0, 0]
+    out[3] .= [1, a[1], b[1]]
+end
+
+h_dense_arrayvec_str = ModelingToolkit.build_function(h_dense_arrayvec, [a, b, c])
+h_dense_arrayvec_oop = eval(h_dense_arrayvec_str[1])
+h_dense_arrayvec_ip! = eval(h_dense_arrayvec_str[2])
+out_1_arrayvec = [Vector{Int64}(undef, 3) for i in 1:3]
+out_2_arrayvec = [Vector{Int64}(undef, 3) for i in 1:3]
+julia_dense_arrayvec = h_dense_arrayvec_julia(input)
+mtk_dense_arrayvec = h_dense_arrayvec_oop(input)
+@test_broken julia_dense_arrayvec == mtk_dense_arrayvec
+mtk_dense_arrayvec = @test_broken h_dense_arrayvec_oop(input)
+h_dense_arrayvec_julia!(out_1_arrayvec, input)
[…]
