+++ Project.toml
@@ -35,6 +35,7 @@ SafeTestsets = "0.0.1"
 SpecialFunctions = "0.7, 0.8, 0.9, 0.10"
 StaticArrays = "0.10, 0.11, 0.12"
 TreeViews = "0.3"
+UnPack = "0.1"
 Unitful = "1.1"
 julia = "1.2"
 
