+++ src/systems/diffeqs/odesystem.jl
@@ -75,15 +75,17 @@ function ODESystem(deqs::AbstractVector{<:Equation}, iv, dvs, ps;
 end
 
 var_from_nested_derivative(x) = var_from_nested_derivative(x,0)
+var_from_nested_derivative(x::Constant) = (missing, missing)
 var_from_nested_derivative(x,i) = x.op isa Differential ? var_from_nested_derivative(x.args[1],i+1) : (x.op,i)
 iv_from_nested_derivative(x) = x.op isa Differential ? iv_from_nested_derivative(x.args[1]) : x.args[1].op
+iv_from_nested_derivative(x::Constant) = missing
 
 function ODESystem(eqs; kwargs...)
-    ivs = unique(iv_from_nested_derivative(eq.lhs) for eq ∈ eqs)
+    ivs = unique(skipmissing(iv_from_nested_derivative(eq.lhs) for eq ∈ eqs))
     length(ivs) == 1 || throw(ArgumentError("one independent variable currently supported"))
     iv = first(ivs)
 
-    dvs = unique(var_from_nested_derivative(eq.lhs)[1] for eq ∈ eqs)
+    dvs = unique(skipmissing(var_from_nested_derivative(eq.lhs)[1] for eq ∈ eqs))
     ps = filter(vars(eq.rhs for eq ∈ eqs)) do x
         isparameter(x) & !isequal(x, iv)
     end |> collect
