### 2020-04-14 d1189ed
Add bigraph conversion and pretty printing for debugging
Co-authored-by: "Shashi Gowda" <gowda@mit.edu>
Co-authored-by: "Yingbo Ma" <mayingbo5@gmail.com>
