### 2020-04-14 6e6cded
Add variable association list
