### 2020-04-14 55dee90
Add Pantelides algorithm
