### 2020-04-14 080c40e
Add maxiter and simplify control flow
