### 2020-04-14 e2df0e8
Add maxiter and error msg
