### 2020-04-14 7ec2a4f
Fix index_reduction tests
