### 2020-04-14 dd89e11
treat symbols symbolically when converting from expressions
