+++ .travis.yml
@@ -20,7 +20,7 @@ codecov: true
 jobs:
   include:
     - stage: "Documentation"
-      julia: 1.4
+      julia: 1.3
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
