### 2020-04-15 1de3e76
Add a manual example of converting graph to eqs
