+++ src/utils.jl
@@ -3,7 +3,7 @@ function Expression(ex;mod=Main)
     ex.head === :if && (ex = Expr(:call, ifelse, ex.args...))
     ex.head === :call || throw(ArgumentError("internal representation does not support non-call Expr"))
 
-    op = getproperty(mod,ex.args[1])
+    op = getproperty(mod,Symbol(ex.args[1]))
     args = convert.(Expression, ex.args[2:end])
 
     return Operation(op, args)
