### 2020-04-16 15d7773
allow dict value maps
