+++ test/lowering_solving.jl
@@ -20,6 +20,8 @@ p  = [σ => 28.0,
       ρ => 10.0,
       β => 8/3]
 
+u0[convert(Variable,D(x))]
+
 tspan = (0.0,100.0)
 prob = ODEProblem(sys,u0,tspan,p,jac=true)
 sol = solve(prob,Tsit5())
