### 2020-04-17 8011bda
Set push_preview=true in makedocs
This allows PRs changing documentation to be previewed ahead of time.
