### 2020-04-17 6a52379
Remove index reduction
