### 2020-04-17 ec1787b
Typo
generate_jacobian takes an ODE and sys is a ::Tuple{ODESystem,Array{Operation,1},Array{Operation,1}}
=> pass sys[1].
