### 2020-04-18 fd22dd8
Update src/variables.jl
Co-Authored-By: Anshul Singhvi <asinghvi17@simons-rock.edu>
