### 2020-04-21 f7a21e8
updating build_function for integrators
