### 2020-04-21 595a530
constant and variable rate assembly working
