+++ .github/workflows/TagBot.yml
@@ -8,4 +8,5 @@ jobs:
     steps:
       - uses: JuliaRegistries/TagBot@v1
         with:
-          token: ${{ secrets.GITHUB_TOKEN }}
+            token: ${{ secrets.GITHUB_TOKEN }}
+            ssh: ${{ secrets.DOCUMENTER_KEY }}
