+++ test/jumpsystem.jl
@@ -1,4 +1,4 @@
-using ModelingToolkit, DiffEqJump, Test, LinearAlgebra
+using ModelingToolkit, DiffEqBase, DiffEqJump, Test, LinearAlgebra
 MT = ModelingToolkit
 
 # basic SIR model with tweaks
@@ -12,7 +12,7 @@ j₁      = ConstantRateJump(rate₁,affect₁)
 j₂      = VariableRateJump(rate₂,affect₂)
 js      = JumpSystem([j₁,j₂], t, [S,I,R], [β,γ])
 mtjump1 = MT.assemble_crj(js, j₁)
-mtjump2 = MT.assemble_crj(js, j₂)
+mtjump2 = MT.assemble_vrj(js, j₂)
 
 # doc version
 rate1(u,p,t) = (0.1/1000.0)*u[1]*u[2]
@@ -31,16 +31,16 @@ jump2 = VariableRateJump(rate2,affect2!)
 # test crjs
 u = [100, 9, 5]
 p = (0.1/1000,0.01)
-t = 1.0
+tf = 1.0
 mutable struct TestInt
     u
     p
     t
 end
-mtintegrator = TestInt(u,p,t)
-integrator   = TestInt(u,p,t)
-@test abs(mtjump1.rate(u,p,t) - jump1.rate(u,p,t)) < 10*eps()
-@test abs(mtjump2.rate(u,p,t) - jump2.rate(u,p,t)) < 10*eps()
+mtintegrator = TestInt(u,p,tf)
+integrator   = TestInt(u,p,tf)
+@test abs(mtjump1.rate(u,p,tf) - jump1.rate(u,p,tf)) < 10*eps()
+@test abs(mtjump2.rate(u,p,tf) - jump2.rate(u,p,tf)) < 10*eps()
 mtjump1.affect!(mtintegrator)
 jump1.affect!(integrator)
 @test norm(integrator.u - mtintegrator.u) < 10*eps()
@@ -50,3 +50,15 @@ jump2.affect!(integrator)
 @test norm(integrator.u - mtintegrator.u) < 10*eps()
 
 
+# test can make and solve a jump problem
+rate₂   = γ*I
+affect₂ = [I ~ I - 1, R ~ R + 1]
+j₃ = ConstantRateJump(rate₂,affect₂)
+js2 = JumpSystem([j₁,j₃], t, [S,I,R], [β,γ])
+u₀ = [999,1,0]; p = (0.1/1000,0.01); tspan = (0.,250.)
+dprob = DiscreteProblem(u₀,tspan,p)
+jprob = JumpProblem(js2, dprob, Direct())
+sol = solve(jprob, SSAStepper())
+
+using Plots
+plot(sol)
