### 2020-04-25 b22a6d5
integrate SymbolicUtils
