+++ src/simplify.jl
@@ -1,97 +1,51 @@
-function simplify_constants(O::Operation, shorten_tree)
-    while true
-        O′ = _simplify_constants(O, shorten_tree)
-        if is_operation(O′)
-            O′ = Operation(O′.op, simplify_constants.(O′.args, shorten_tree))
-        end
-        isequal(O, O′) && return O
-        O = O′
+import SymbolicUtils
+import SymbolicUtils: FnType
+
+# ModelingToolkit -> SymbolicUtils
+SymbolicUtils.istree(x::Operation) = true
+function SymbolicUtils.operation(x::Operation)
+    if x.op isa Variable
+        T = FnType{NTuple{length(x.args), Any}, vartype(x.op)}
+        SymbolicUtils.Variable{T}(x.op.name)
+    else
+        x.op
     end
 end
-simplify_constants(x, shorten_tree) = x
 
-"""
-simplify_constants(x::Operation)
+# This is required to infer the right type for
+# Operation(Variable{Parameter{Number}}(:foo), [])
+# While keeping the metadata that the variable is a parameter.
+SymbolicUtils.promote_symtype(f::SymbolicUtils.Variable{FnType{X,Parameter{Y}}},
+                              xs...) where {X, Y} = Y
 
-Simplifies the constants within an expression, for example removing equations
-multiplied by a zero and summing constant values.
-"""
-simplify_constants(x) = simplify_constants(x, true)
+SymbolicUtils.arguments(x::Operation) = x.args
 
-Base.isone(x::Operation) = x.op == one || x.op == Constant && isone(x.args)
-const AC_OPERATORS = (*, +)
+# SymbolicUtils wants raw numbers
+SymbolicUtils.to_symbolic(x::Constant) = x.value
+SymbolicUtils.to_symbolic(x::Variable{T}) where {T} = SymbolicUtils.Variable{T}(x.name)
 
-function _simplify_constants(O::Operation, shorten_tree)
-    # Tree shrinking
-    if shorten_tree && O.op ∈ AC_OPERATORS
-        # Flatten tree
-        idxs = findall(x -> is_operation(x) && x.op === O.op, O.args)
-        if !isempty(idxs)
-            keep_idxs = eachindex(O.args) .∉ (idxs,)
-            args = Vector{Expression}[O.args[i].args for i in idxs]
-            push!(args, O.args[keep_idxs])
-            return Operation(O.op, vcat(args...))
-        end
+# Optional types of vars
+# Once converted to SymbolicUtils Variable, a Parameter needs to hide its metadata
+_vartype(x::Variable{<:Parameter{T}}) where {T} = T
+_vartype(x::Variable{T}) where {T} = T
+SymbolicUtils.symtype(x::Variable) = _vartype(x) # needed for a()
+SymbolicUtils.symtype(x::SymbolicUtils.Variable{<:Parameter{T}}) where {T} = T
 
-        # Collapse constants
-        idxs = findall(is_constant, O.args)
-        if length(idxs) > 1
-            other_idxs = eachindex(O.args) .∉ (idxs,)
-            new_const = Constant(mapreduce(get, O.op, O.args[idxs]))
-            args = push!(O.args[other_idxs], new_const)
+# returning Any causes SymbolicUtils to infer the type using `promote_symtype`
+# But we are OK with Number here for now I guess
+SymbolicUtils.symtype(x::Expression) = Number
 
-            length(args) == 1 && return first(args)
-            return Operation(O.op, args)
-        end
-    end
 
-    if O.op === (*)
-        # If any variable is `Constant(0)`, zero the whole thing
-        any(iszero, O.args) && return Constant(0)
+# SymbolicUtils -> ModelingToolkit
 
[…]
