### 2020-04-25 128f737
sometimes Number sometimes Constant
