+++ src/differentials.jl
@@ -52,7 +52,7 @@ function expand_derivatives(O::Operation)
         end |> simplify_constants
     end
 
-    return O
+    return O |> simplify_constants
 end
 expand_derivatives(x) = x
 
