+++ test/derivatives.jl
@@ -15,12 +15,12 @@ dsin = D(sin(t))
 @test isequal(expand_derivatives(dsin), cos(t))
 
 dcsch = D(csch(t))
-@test isequal(expand_derivatives(dcsch), simplify_constants(coth(t) * csch(t) * -1))
+@test isequal(expand_derivatives(dcsch), simplify_constants(-coth(t) * csch(t)))
 
 @test isequal(expand_derivatives(D(-7)), 0)
 @test isequal(expand_derivatives(D(sin(2t))), simplify_constants(cos(2t) * 2))
 @test isequal(expand_derivatives(D2(sin(t))), simplify_constants(-sin(t)))
-@test isequal(expand_derivatives(D2(sin(2t))), simplify_constants(sin(2t) * -4))
+@test isequal(expand_derivatives(D2(sin(2t))), simplify_constants(-sin(2t) * 4))
 @test isequal(expand_derivatives(D2(t)), 0)
 @test isequal(expand_derivatives(D2(5)), 0)
 
@@ -30,20 +30,20 @@ dsinsin = D(sin(sin(t)))
 
 d1 = D(sin(t)*t)
 d2 = D(sin(t)*cos(t))
-@test isequal(expand_derivatives(d1), t*cos(t)+sin(t))
-@test isequal(expand_derivatives(d2), simplify_constants(cos(t)*cos(t)+(sin(t)*-1)*sin(t)))
+@test isequal(expand_derivatives(d1), simplify_constants(t*cos(t)+sin(t)))
+@test isequal(expand_derivatives(d2), simplify_constants(cos(t)*cos(t)+(-sin(t))*sin(t)))
 
 eqs = [0 ~ σ*(y-x),
        0 ~ x*(ρ-z)-y,
        0 ~ x*y - β*z]
 sys = NonlinearSystem(eqs, [x,y,z], [σ,ρ,β])
 jac = calculate_jacobian(sys)
-@test isequal(jac[1,1], σ*-1)
+@test isequal(jac[1,1], -1σ)
 @test isequal(jac[1,2], σ)
 @test isequal(jac[1,3], 0)
-@test isequal(jac[2,1], ρ-z)
+@test isequal(jac[2,1], -1z + ρ) # FIXME
 @test isequal(jac[2,2], -1)
-@test isequal(jac[2,3], x*-1)
+@test isequal(jac[2,3], -1x)
 @test isequal(jac[3,1], y)
 @test isequal(jac[3,2], x)
 @test isequal(jac[3,3], -1*β)
@@ -57,7 +57,7 @@ jac = calculate_jacobian(sys)
 @variables x(t) y(t) z(t)
 
 @test isequal(expand_derivatives(D(x * y)), simplify_constants(y*D(x) + x*D(y)))
-@test_broken isequal(expand_derivatives(D(x * y)), simplify_constants(D(x)*y + x*D(y)))
+@test isequal(expand_derivatives(D(x * y)), simplify_constants(D(x)*y + x*D(y)))
 
 @test isequal(expand_derivatives(D(2t)), 2)
 @test isequal(expand_derivatives(D(2x)), 2D(x))
