### 2020-04-25 f1cc897
assume simplify always returns Expression
