### 2020-04-29 c12eec2
lower bound SymbolicUtils
