### 2020-04-29 69690e5
fixed rho-z needing to be -z + rho
