+++ test/reactionsystem.jl
@@ -1,11 +1,11 @@
 using ModelingToolkit, LinearAlgebra, Test
 
-@parameters t k[1:13]
+@parameters t k[1:15]
 @variables A(t) B(t) C(t) D(t)
 rxs = [Reaction(k[1], nothing, [A]),            # 0 -> A
        Reaction(k[2], [B], nothing),            # B -> 0
        Reaction(k[3],[A],[C]),                  # A -> C
-       Reaction(k[4], [C], [A,B]),              # C -> A + B        
+       Reaction(k[4], [C], [A,B]),              # C -> A + B
        Reaction(k[5], [C], [A], [1], [2]),      # C -> A + A
        Reaction(k[6], [A,B], [C]),              # A + B -> C
        Reaction(k[7], [B], [A], [2], [1]),      # 2B -> A
@@ -14,8 +14,10 @@ rxs = [Reaction(k[1], nothing, [A]),            # 0 -> A
        Reaction(k[10], [A], [C,D], [2], [1,1]), # 2A -> C + D
        Reaction(k[11], [A], [A,B], [2], [1,1]), # 2A -> A + B
        Reaction(k[12], [A,B,C], [C,D], [1,3,4], [2, 3]), # A+3B+4C -> 2C + 3D
-       Reaction(k[13]*A/(2+A), [A], nothing; only_use_rate=true)  # A -> 0 with custom rate 
-       ]  
+       Reaction(k[13], [A,B], nothing, [2,1], nothing), # 2A+B -> 0
+       Reaction(k[14], nothing, [A], nothing, [3]), # 0 -> 3A
+       Reaction(k[15]*A/(2+A), [A], nothing; only_use_rate=true)  # A -> 0 with custom rate
+       ]
 rs = ReactionSystem(rxs,t,[A,B,C,D],k)
 odesys = convert(ODESystem,rs)
 sdesys = convert(SDESystem,rs)
@@ -50,7 +52,7 @@ function sdenoise(u,k,t)
             -sqrt(k[11]*A^2/2) sqrt(k[11]*A^2/2) z z;
             -sqrt(k[12]*A*B^3*C^4/144) -3*sqrt(k[12]*A*B^3*C^4/144) -2*sqrt(k[12]*A*B^3*C^4/144) 3*sqrt(k[12]*A*B^3*C^4/144);
             -sqrt(k[13]*A/(2+A)) z z z]'
-       
+
        return G
 end
 
