### 2020-05-02 7cb91dc
- Throws warning when both substrate and product vector is nothing. - Throws warning when someone gives stochiometry for a 'nothing" substrate or product input. - Empty substrate/product stoichimoetry vectors take the type of the other.
