### 2020-05-02 9aabe56
- Added warning for when both stoichiometry vectors are nothing.
