### 2020-05-02 e71c0ec
- Updated test systems functions to take the two added test reactions into account.
