### 2020-05-03 2229f3e
Fix parsing of variable names when gensym'ed
