### 2020-05-03 14d2551
unpack arrays into tuples
