### 2020-05-03 bc73326
fix SDEProblem construction
