### 2020-05-03 07f6e2d
- Remove unnecessary check.
