### 2020-05-03 1d71d7a
add outputidxs
