### 2020-05-04 3968efd
add test for generated jump system
