### 2020-05-04 7c1172e
add DiscreteProblem with u0/param maps
