+++ src/ModelingToolkit.jl
@@ -103,7 +103,8 @@ include("build_function.jl")
 export ODESystem, ODEFunction
 export SDESystem, SDEFunction
 export JumpSystem
-export ODEProblem, SDEProblem, NonlinearProblem, OptimizationProblem, JumpProblem
+export ODEProblem, SDEProblem, NonlinearProblem, OptimizationProblem
+export JumpProblem, DiscreteProblem
 export NonlinearSystem, OptimizationSystem
 export ode_order_lowering
 export PDESystem
