### 2020-05-05 c42b3e4
add mass action jump support
