+++ test/jumpsystem.jl
@@ -106,3 +106,18 @@ m2 = getmean(jprob,Nsims)
 
 # test JumpSystem solution agrees with direct version
 @test abs(m-m2) ./ m < .01
+
+
+# mass action jump tests for SIR model
+maj1 = MassActionJump(2*β/2, [S => 1, I => 1], [S => -1, I => 1])
+maj2 = MassActionJump(γ, [I => 1], [I => -1, R => 1])
+js   = JumpSystem([maj1,maj2], t, [S,I,R], [β,γ])
+statetoid = Dict(convert(Variable,state) => i for (i,state) in enumerate(states(js)))
+ptoid     = Dict(convert(Variable,par) => i for (i,par) in enumerate(parameters(js)))
+dprob = DiscreteProblem(js, u₀map, tspan, parammap)
+jprob = JumpProblem(js, dprob, Direct())
+m3 = getmean(jprob,Nsims)
+@test abs(m2-m3) < .01
+
+# mass action jump tests for other reaction types (zero order, second order, decay)
+# TODO
