### 2020-05-06 ad15802
substitute using SymbolicUtils
Fixes https://github.com/SciML/ModelingToolkit.jl/issues/339

MWE:

```julia
using ModelingToolkit
@parameters α β γ δ
s = α => γ
expr = (((1 / β - 1) + δ) / α) ^ (1 / (α - 1))
ModelingToolkit.substitute(expr, s)
new  = (((1 / β - 1) + δ) / γ) ^ (1 / (γ - 1))
isequal(expr, new)
```
