+++ src/systems/jumps/jumpsystem.jl
@@ -9,24 +9,24 @@ struct JumpSystem <: AbstractSystem
     systems::Vector{JumpSystem}
 end
 
-function JumpSystem(eqs, iv, states, ps; systems = JumpSystem[], 
+function JumpSystem(eqs, iv, states, ps; systems = JumpSystem[],
                                           name = gensym(:JumpSystem))
     JumpSystem(eqs, iv, convert.(Variable, states), convert.(Variable, ps), name, systems)
 end
 
 
 
-generate_rate_function(js, rate) = build_function(rate, states(js), parameters(js), 
-                                                  independent_variable(js), 
+generate_rate_function(js, rate) = build_function(rate, states(js), parameters(js),
+                                                  independent_variable(js),
                                                   expression=Val{false})
 
-generate_affect_function(js, affect, outputidxs) = build_function(affect, states(js), 
-                                                      parameters(js), 
+generate_affect_function(js, affect, outputidxs) = build_function(affect, states(js),
+                                                      parameters(js),
                                                       independent_variable(js),
                                                       expression=Val{false},
-                                                      headerfun=add_integrator_header, 
+                                                      headerfun=add_integrator_header,
                                                       outputidxs=outputidxs)[2]
-function assemble_vrj(js, vrj, statetoid) 
+function assemble_vrj(js, vrj, statetoid)
     rate   = generate_rate_function(js, vrj.rate)
     outputvars = (convert(Variable,affect.lhs) for affect in vrj.affect!)
     outputidxs = ((statetoid[var] for var in outputvars)...,)
@@ -42,18 +42,20 @@ function assemble_crj(js, crj, statetoid)
     ConstantRateJump(rate, affect)
 end
 
-function assemble_maj(js, maj::MassActionJump{U,Vector{Pair{V,W}},Vector{Pair{V2,W2}}}, 
-                      statetoid, ptoid, p, pcontext) where {U,V,W,V2,W2}
+function assemble_maj(js, maj::MassActionJump{U,Vector{Pair{V,W}},Vector{Pair{V2,W2}}},
+                      statetoid, ptoid, parammap) where {U,V,W,V2,W2}
     sr = maj.scaled_rates
-    if sr isa Operation || sr isa Variable
-        pval = Base.eval(pcontext, Expr(maj.scaled_rates))
-    else   
+    if sr isa Operation
+        pval = substitute(sr,parammap)
+    elseif sr isa Variable
+        pval = Dict(parammap)[sr()]
+    else
         pval = maj.scaled_rates
     end
-    
+
     rs = Vector{Pair{Int,W}}()
     for (spec,stoich) in maj.reactant_stoch
-        if iszero(spec) 
+        if iszero(spec)
             push!(rs, 0 => stoich)
         else
             push!(rs, statetoid[convert(Variable,spec)] => stoich)
@@ -73,13 +75,13 @@ end
 
 """
 ```julia
-function DiffEqBase.DiscreteProblem(sys::AbstractSystem, u0map, tspan, 
+function DiffEqBase.DiscreteProblem(sys::AbstractSystem, u0map, tspan,
                                     parammap=DiffEqBase.NullParameters; kwargs...)
 ```
 
 Generates a DiscreteProblem from an AbstractSystem
 """
-function DiffEqBase.DiscreteProblem(sys::AbstractSystem, u0map, tspan::Tuple, 
+function DiffEqBase.DiscreteProblem(sys::AbstractSystem, u0map, tspan::Tuple,
                                     parammap=DiffEqBase.NullParameters(); kwargs...)
[…]
