### 2020-05-06 afab565
allow more types in substitution, refactor a bit
