### 2020-05-06 ba14aed
use substitute function from SymbolicUtils
