### 2020-05-07 a35b535
don't extend SymbolicUtils.simplify
