### 2020-05-07 b018411
simplify stuff to Constant
