### 2020-05-07 108593e
try and fix latexificaition with components
https://github.com/SciML/ModelingToolkit.jl/issues/305
