+++ Project.toml
@@ -37,7 +37,7 @@ NaNMath = "0.3"
 SafeTestsets = "0.0.1"
 SpecialFunctions = "0.7, 0.8, 0.9, 0.10"
 StaticArrays = "0.10, 0.11, 0.12"
-SymbolicUtils = "0.2.1"
+SymbolicUtils = "0.3"
 TreeViews = "0.3"
 UnPack = "0.1"
 Unitful = "1.1"
