### 2020-05-07 645d8a3
CompatHelper: bump compat for "SymbolicUtils" to "0.3"
