+++ Project.toml
@@ -39,7 +39,7 @@ SpecialFunctions = "0.7, 0.8, 0.9, 0.10"
 StaticArrays = "0.10, 0.11, 0.12"
 SymbolicUtils = "0.1.1, 0.2"
 TreeViews = "0.3"
-UnPack = "0.1"
+UnPack = "0.1, 1.0"
 Unitful = "1.1"
 julia = "1.2"
 
