### 2020-05-07 2d6d4c3
- JumpSystem is now properly created for Variable and Constant RateJumps. - Version of oderatelawy, jumpratelaw function (possibly unnecessary). - Added var2op utility function (possibly unnecessary).
