### 2020-05-08 b475d63
- MassActionJump not used when only_use_rate=true.
