+++ src/systems/reaction/reactionsystem.jl
@@ -135,8 +135,8 @@ function assemble_jumps(rs)
         elseif rx.only_use_rate || any([isequal(state,r_op) for state in rs.states, r_op in getfield.(get_variables(rx.rate),:op)])
             push!(eqs,ConstantRateJump(rl,affect))
         else
-            reactant_stoch = Pair.(var2op.(getfield.(rx.substrates,:op)),rx.substoich)
-            net_stoch = isempty(rx.substoich) ? [0 => 1]  : map(p -> Pair(var2op(p[1]),p[2]),rx.netstoich)
+            reactant_stoch = isempty(rx.substoich) ? [0 => 1] : Pair.(var2op.(getfield.(rx.substrates,:op)),rx.substoich)
+            net_stoch = map(p -> Pair(var2op(p[1]),p[2]),rx.netstoich)
             push!(eqs,MassActionJump(rx.rate, reactant_stoch, net_stoch))
         end
     end
