### 2020-05-08 5e44464
don't call to_mtk after calling ModelingToolkit.simplify
