+++ src/build_function.jl
@@ -4,6 +4,11 @@ struct StanTarget <: BuildTargets end
 struct CTarget <: BuildTargets end
 struct MATLABTarget <: BuildTargets end
 
+abstract type ParallelForm end
+struct SerialForm <: ParallelForm end
+struct MultithreadedForm <: ParallelForm end
+struct DistributedForm <: ParallelForm end
+
 """
 `build_function`
 
@@ -57,7 +62,7 @@ function build_function(args...;target = JuliaTarget(),kwargs...)
 end
 
 function addheader(ex, fargs, iip; X=gensym(:MTIIPVar))
-  if iip  
+  if iip
     wrappedex = :(
         ($X,$(fargs.args...)) -> begin
         $ex
@@ -69,16 +74,16 @@ function addheader(ex, fargs, iip; X=gensym(:MTIIPVar))
       ($(fargs.args...),) -> begin
         $ex
       end
-    )  
+    )
   end
   wrappedex
 end
 
 function add_integrator_header(ex, fargs, iip; X=gensym(:MTIIPVar))
   integrator = gensym(:MTKIntegrator)
-  if iip  
+  if iip
     wrappedex = :(
-        $integrator -> begin 
+        $integrator -> begin
         ($X,$(fargs.args...)) = (($integrator).u,($integrator).u,($integrator).p,($integrator).t)
         $ex
         nothing
@@ -86,11 +91,11 @@ function add_integrator_header(ex, fargs, iip; X=gensym(:MTIIPVar))
     )
   else
     wrappedex = :(
-      $integrator -> begin 
+      $integrator -> begin
       ($(fargs.args...),) = (($integrator).u,($integrator).p,($integrator).t)
         $ex
       end
-    )  
+    )
   end
   wrappedex
 end
@@ -114,7 +119,7 @@ function _build_function(target::JuliaTarget, op::Operation, args...;
 
     fargs = Expr(:tuple,argnames...)
     oop_ex = headerfun(bounds_block, fargs, false)
-        
+
     if !linenumbers
         oop_ex = striplines(oop_ex)
     end
@@ -129,8 +134,15 @@ end
 function _build_function(target::JuliaTarget, rhss, args...;
                          conv = simplified_expr, expression = Val{true},
                          checkbounds = false, constructor=nothing,
-                         linenumbers = false, multithread=false, 
-                         headerfun=addheader, outputidxs=nothing)
+                         linenumbers = false, multithread=nothing,
+                         headerfun=addheader, outputidxs=nothing,
+						 parallel=SerialForm())
+
+	if multithread isa Bool
+		@warn("multithraded is deprecated for the parallel argument. See the documentation.")
+		parallel = multithread ? MultithreadedForm() : SerialForm()
+	end
+
     argnames = [gensym(:MTKArg) for i in 1:length(args)]
     arg_pairs = map(vars_to_pairs,zip(argnames,args))
     ls = reduce(vcat,first.(arg_pairs))
@@ -143,23 +155,39 @@ function _build_function(target::JuliaTarget, rhss, args...;
 
[…]
