### 2020-05-08 7072cfc
push simplify through derivative expansion
