+++ src/direct.jl
@@ -7,8 +7,7 @@ A helper function for computing the gradient of an expression with respect to
 an array of variable expressions.
 """
 function gradient(O::Expression, vars::AbstractVector{<:Expression}; simplify = true)
-    out = [expand_derivatives(Differential(v)(O)) for v in vars]
-    simplify ? ModelingToolkit.simplify.(out) : out
+    [expand_derivatives(Differential(v)(O),simplify) for v in vars]
 end
 
 """
@@ -20,8 +19,7 @@ A helper function for computing the Jacobian of an array of expressions with res
 an array of variable expressions.
 """
 function jacobian(ops::AbstractVector{<:Expression}, vars::AbstractVector{<:Expression}; simplify = true)
-    out = [expand_derivatives(Differential(v)(O)) for O in ops, v in vars]
-    simplify ? ModelingToolkit.simplify.(out) : out
+    [expand_derivatives(Differential(v)(O),simplify) for O in ops, v in vars]
 end
 
 """
@@ -33,8 +31,7 @@ A helper function for computing the Hessian of an expression with respect to
 an array of variable expressions.
 """
 function hessian(O::Expression, vars::AbstractVector{<:Expression}; simplify = true)
-    out = [expand_derivatives(Differential(v2)(Differential(v1)(O))) for v1 in vars, v2 in vars]
-    simplify ? ModelingToolkit.simplify.(out) : out
+    [expand_derivatives(Differential(v2)(Differential(v1)(O)),simplify) for v1 in vars, v2 in vars]
 end
 
 function simplified_expr(O::Operation)
