+++ test/bigsystem.jl
@@ -45,13 +45,31 @@ end
 
 f(du,u,nothing,0.0)
 
-multithreadedf = eval(ModelingToolkit.build_function(du,u,parallel=MultithreadedForm())[2])
+multithreadedf = eval(ModelingToolkit.build_function(du,u,parallel=ModelingToolkit.MultithreadedForm())[2])
 _du = rand(N,N,3)
 _u = rand(N,N,3)
 multithreadedf(_du,_u)
 
-jac = sparse(ModelingToolkit.jacobian(vec(du),vec(u),simplify=false))
-multithreadedjac = eval(ModelingToolkit.build_function(vec(jac),u,parallel=MultithreadedForm())[2])
+using Distributed
+addprocs(4)
+distributedf = eval(ModelingToolkit.build_function(du,u,parallel=ModelingToolkit.DistributedForm())[2])
 
-#_jac = similar(jac,Float64)
-#multithreadedjac(_jac,_u)
+jac = sparse(ModelingToolkit.jacobian(vec(du),vec(u),simplify=false))
+serialjac = eval(ModelingToolkit.build_function(vec(jac),u)[2])
+multithreadedjac = eval(ModelingToolkit.build_function(vec(jac),u,parallel=ModelingToolkit.MultithreadedForm())[2])
+distributedjac = eval(ModelingToolkit.build_function(vec(jac),u,parallel=ModelingToolkit.DistributedForm())[2])
+
+#=
+MyA = zeros(N,N)
+AMx = zeros(N,N)
+DA  = zeros(N,N)
+using BenchmarkTools
+@btime f(_du,_u,nothing,0.0)
+@btime multithreadedf(_du,_u)
+@btime distributedf(_du,_u)
+
+_jac = similar(jac,Float64)
+@btime serialjac(_jac,_u)
+@btime multithreadedjac(_jac,_u)
+@btime distributedjac(_jac,_u)
+=#
