### 2020-05-08 b3f3bd1
update to reusable closures and fix tests
