### 2020-05-08 11b9128
let SymbolicUtils figure out types of expressions
