### 2020-05-08 319873d
actually do the conversion -- it was going into simplify(::Any) because arg is a symutils expression
