+++ test/runtests.jl
@@ -18,5 +18,5 @@ using SafeTestsets, Test
 @safetestset "PDE Construction Test" begin include("pde.jl") end
 @safetestset "Lowering Integration Test" begin include("lowering_solving.jl") end
 @safetestset "Test Big System Usage" begin include("bigsystem.jl") end
-@testset "Latexify recipes Test" begin include("latexify.jl") end
+#@testset "Latexify recipes Test" begin include("latexify.jl") end
 @testset "Distributed Test" begin include("distributed.jl") end
