### 2020-05-09 15700e5
remove distributed from Travis
