### 2020-05-09 378eb49
keep the same result, do just 1 simplification
