### 2020-05-11 421a616
fix mrj substitution to rate constant
