+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v3.3",
   "v3.2",
   "v3.1",
   "dev",
