### 2020-05-11 c6b2653
- Added another few test reactions (to enable proper testing of Constant and Variable RateJump's.
