### 2020-05-11 0a9e73a
- Added another few test reactions (to enable proper testing of Constant and Variable RateJump's (all changes).
