### 2020-05-12 a81446e
store js eqs as arraypartition
