### 2020-05-12 5bbda08
pass syms to DiscreteProblem for plotting
