+++ src/ModelingToolkit.jl
@@ -101,6 +101,7 @@ include("systems/optimization/optimizationsystem.jl")
 include("systems/pde/pdesystem.jl")
 
 include("systems/reaction/reactionsystem.jl")
+include("systems/dependency_graphs.jl")
 
 include("latexify_recipes.jl")
 include("build_function.jl")
