+++ src/ModelingToolkit.jl
@@ -120,7 +120,6 @@ export IntervalDomain, ProductDomain, ⊗, CircleDomain
 export Equation, ConstrainedEquation
 export Operation, Expression, Variable
 export independent_variable, states, parameters, equations 
-export numstates, numparameters, numequations
 
 export calculate_jacobian, generate_jacobian, generate_function
 export calculate_tgrad, generate_tgrad
