### 2020-05-14 1768e27
add variable dependencies
