+++ src/systems/dependency_graphs.jl
@@ -6,7 +6,7 @@ function equation_dependencies(sys::AbstractSystem; variables=states(sys))
     depeqs_to_vars = Vector{Vector{Variable}}(undef,length(eqs))
 
     for (i,eq) in enumerate(eqs)      
-        depeqs_to_vars[i] = collect(equation_dependencies!(deps, eq, variables))
+        depeqs_to_vars[i] = collect(get_variables!(deps, eq, variables))
         empty!(deps)
     end
 
@@ -31,7 +31,7 @@ function asgraph(eqdeps, vtois)
     ne = 0
     for (eqidx,vidxs) in enumerate(fadjlist)
         foreach(vidx -> push!(badjlist[vidx], eqidx), vidxs)
-        ne += 1
+        ne += length(vidxs)
     end
 
     BipartiteGraph(ne, fadjlist, badjlist)
@@ -39,8 +39,33 @@ end
 
 # could be made to directly generate graph and save memory
 function asgraph(sys::AbstractSystem; variables=nothing, variablestoids=nothing)
-    vs   = isnothing(variables) ? states(sys) : variables    
+    vs     = isnothing(variables) ? states(sys) : variables    
     eqdeps = equation_dependencies(sys, variables=vs)
     vtois  = isnothing(variablestoids) ? Dict(convert(Variable, v) => i for (i,v) in enumerate(vs)) : variablestoids
     asgraph(eqdeps, vtois)
+end
+
+
+# for each variable determine the equations that modify it
+function variable_dependencies(sys::AbstractSystem; variables=states(sys), variablestoids=nothing)
+    eqs   = equations(sys)
+    vtois = isnothing(variablestoids) ? Dict(convert(Variable, v) => i for (i,v) in enumerate(variables)) : variablestoids
+
+    deps = Set{Variable}()
+    badjlist = Vector{Vector{Int}}(undef, length(eqs))    
+    for (eidx,eq) in enumerate(eqs)
+        modified_states!(deps, eq, variables)
+        badjlist[eidx] = sort!([vtois[var] for var in deps])
+        empty!(deps)
+    end
+
+    fadjlist = [Vector{Int}() for i = 1:length(variables)]
+    ne = 0
+    for (eqidx,vidxs) in enumerate(badjlist)
+        println(vidxs)
+        foreach(vidx -> push!(fadjlist[vidx], eqidx), vidxs)
+        ne += length(vidxs)
+    end
+
+    BipartiteGraph(ne, fadjlist, badjlist)
 end
