### 2020-05-14 909c6a8
add conversion to LightGraph.SimpleDiGraph
