+++ src/systems/dependency_graphs.jl
@@ -45,7 +45,6 @@ function asgraph(sys::AbstractSystem; variables=nothing, variablestoids=nothing)
     asgraph(eqdeps, vtois)
 end
 
-
 # for each variable determine the equations that modify it
 function variable_dependencies(sys::AbstractSystem; variables=states(sys), variablestoids=nothing)
     eqs   = equations(sys)
@@ -62,10 +61,43 @@ function variable_dependencies(sys::AbstractSystem; variables=states(sys), varia
     fadjlist = [Vector{Int}() for i = 1:length(variables)]
     ne = 0
     for (eqidx,vidxs) in enumerate(badjlist)
-        println(vidxs)
         foreach(vidx -> push!(fadjlist[vidx], eqidx), vidxs)
         ne += length(vidxs)
     end
 
     BipartiteGraph(ne, fadjlist, badjlist)
+end
+
+# convert BipartiteGraph to LightGraph.SimpleDiGraph
+function digraph(g::BipartiteGraph, sys::AbstractSystem; variables = states(sys), equationsfirst = true)
+    neqs  = length(equations(sys))
+    nvars = length(variables)
+
+    fadjlist = deepcopy(g.fadjlist)
+    badjlist = deepcopy(g.badjlist)
+    if equationsfirst
+        # variable indices must be incremented by neqs
+        for i = 1:neqs
+            fadjlist[i] .+= neqs
+        end
+
+        # variables do not connect to anything
+        append!(fadjlist, [Vector{Int}() for i=1:nvars])
+
+        # eqs have nothing that mapped to them
+        prepend!(badjlist, [Vector{Int}() for i=1:neqs])
+    else
+        # equation indices must be incremented by nvars
+        for i = 1:nvars
+            fadjlist[i] .+= nvars
+        end
+
+        # equations do not connect to anything
+        append!(fadjlist, [Vector{Int}() for i=1:neqs])
+
+        # vars have nothing that mapped to them
+        prepend!(badjlist, [Vector{Int}() for i=1:nvars])
+    end
+
+    SimpleDiGraph(g.ne, fadjlist, badjlist)
 end
