### 2020-05-14 a0d22c2
cleanup digraph generation
