+++ src/ModelingToolkit.jl
@@ -18,7 +18,7 @@ using RecursiveArrayTools
 import SymbolicUtils
 import SymbolicUtils: to_symbolic, FnType
 
-import LightGraphs: SimpleDiGraph
+import LightGraphs: SimpleDiGraph, add_edge!
 
 import TreeViews
 
@@ -129,6 +129,9 @@ export calculate_gradient, generate_gradient
 export calculate_factorized_W, generate_factorized_W
 export calculate_hessian, generate_hessian
 export calculate_massmatrix, generate_diffusion_function
+
+export BipartiteGraph, equation_dependencies, variable_dependencies, eqeq_dendencies, varvar_dependencies
+export asgraph, asdigraph
 
 export simplified_expr, rename, get_variables
 export simplify, substitute
