### 2020-05-15 ca9b78f
simplify varvar_deps
