+++ test/runtests.jl
@@ -19,5 +19,6 @@ using SafeTestsets, Test
 @safetestset "PDE Construction Test" begin include("pde.jl") end
 @safetestset "Lowering Integration Test" begin include("lowering_solving.jl") end
 @safetestset "Test Big System Usage" begin include("bigsystem.jl") end
+@safetestset "Depdendency Graph Test" begin include("dep_graphs.jl")
 #@testset "Latexify recipes Test" begin include("latexify.jl") end
 @testset "Distributed Test" begin include("distributed.jl") end
