+++ test/jumpsystem.jl
@@ -118,6 +118,16 @@ jprob = JumpProblem(js3, dprob, Direct())
 m3 = getmean(jprob,Nsims)
 @test abs(m-m3)/m < .01
 
+# maj jump test with dep graphs
+js3b = JumpSystem([maj1,maj2], t, [S,I,R], [β,γ])
+jprobb = JumpProblem(js3b, dprob, RSSA())
+#m4 = getmean(jprobb,Nsims)
+#@test abs(m-m4)/m < .01
+
+jprobc = JumpProblem(js3b, dprob, DirectCR())
+#m5 = getmean(jprobc,Nsims)
+#@test abs(m-m5)/m < .01
+
 # mass action jump tests for other reaction types (zero order, decay)
 maj1 = MassActionJump(2.0, [0 => 1], [S => 1])
 maj2 = MassActionJump(γ, [S => 1], [S => -1])
@@ -138,3 +148,4 @@ ptoid     = Dict(convert(Variable,par) => i for (i,par) in enumerate(parameters(
 dprob = DiscreteProblem(js4, [S => 999], (0,1000.), [β => 100.,γ => .01])
 jprob = JumpProblem(js4, dprob, Direct())
 sol = solve(jprob, SSAStepper());
+
