### 2020-05-15 eb5089b
remove arraypartition slicing in ReactionSystem tests
