### 2020-05-15 1727b3b
reenable type test
