### 2020-05-15 b16a989
remove crashing SSAs in tests
