+++ test/jumpsystem.jl
@@ -120,13 +120,9 @@ m3 = getmean(jprob,Nsims)
 
 # maj jump test with dep graphs
 js3b = JumpSystem([maj1,maj2], t, [S,I,R], [β,γ])
-jprobb = JumpProblem(js3b, dprob, RSSA())
-#m4 = getmean(jprobb,Nsims)
-#@test abs(m-m4)/m < .01
-
-jprobc = JumpProblem(js3b, dprob, DirectCR())
-#m5 = getmean(jprobc,Nsims)
-#@test abs(m-m5)/m < .01
+jprobb = JumpProblem(js3b, dprob, NRM())
+m4 = getmean(jprobb,Nsims)
+@test abs(m-m4)/m < .01
 
 # mass action jump tests for other reaction types (zero order, decay)
 maj1 = MassActionJump(2.0, [0 => 1], [S => 1])
