### 2020-05-16 69de741
CompatHelper: add new compat entry for "LightGraphs" at version "1.3"
