+++ Project.toml
@@ -34,6 +34,7 @@ DiffRules = "0.1, 1.0"
 DocStringExtensions = "0.7, 0.8"
 GeneralizedGenerated = "0.1.4, 0.2"
 Latexify = "0.11, 0.12, 0.13"
+LightGraphs = "1.3"
 MacroTools = "0.5"
 NaNMath = "0.3"
 RecursiveArrayTools = "2.3"
