### 2020-05-17 5518c89
add ismassaction and get rid of some temp arrays
