### 2020-05-17 670c3aa
shorten ismassaction a bit
