+++ src/systems/reaction/reactionsystem.jl
@@ -143,7 +143,7 @@ end
 # if haveivdep=false then time dependent rates will still be classified as mass action
 function ismassaction(rx, rs; rxvars = get_variables(rx.rate), 
                               haveivdep = any(var -> isequal(rs.iv,convert(Variable,var)), rxvars))    
-    return !(haveivdep || rx.only_use_rate || any((isequal(state,convert(Variable,rxv)) for state in states(rs), rxv in rxvars)))
+    return !(haveivdep || rx.only_use_rate || any(convert(Variable,rxv) in states(rs) for rxv in rxvars))
 end
 
 function assemble_jumps(rs)
