+++ src/systems/reaction/reactionsystem.jl
@@ -154,7 +154,7 @@ function assemble_jumps(rs)
         haveivdep = any(var -> isequal(rs.iv,convert(Variable,var)), rxvars)
         if ismassaction(rx, rs; rxvars=rxvars, haveivdep=haveivdep)            
             reactant_stoch = isempty(rx.substoich) ? [0 => 1] : [var2op(sub.op) => stoich for (sub,stoich) in zip(rx.substrates,rx.substoich)]
-            net_stoch      = map(p -> Pair(var2op(p[1]),p[2]), rx.netstoich)
+            net_stoch      = [Pair(var2op(p[1]),p[2]) for p in rx.netstoich]
             push!(eqs, MassActionJump(rx.rate, reactant_stoch, net_stoch))
         else
             rl     = jumpratelaw(rx, rxvars=rxvars)
