+++ v3.4.0/siteinfo.js
@@ -1,0 +1,1 @@
+var DOCUMENTER_CURRENT_VERSION = "v3.4.0";
