### 2020-05-18 33472bb
fix modelingtoolkize to only return the system
This is somewhat of a funny case, because we document it as only returning the system, but it was still doing the old thing of returning the states and parameters as well (which is redundant). So I'm going to call it not breaking as we changed the docs awhile ago, so it's moreso just a weird error that the package was updated after the docs.
